### 2022-05-11 #534 @msperlin
Download and Parse Public Data Released by B3 Exchange
Submitting Author Name: Marcelo S. Perlin
Submitting Author Github Handle: @msperlin
Repository:  https://github.com/wilsonfreitas/rb3
Version submitted: 0.0.3
Submission type: Standard
Version accepted: TBD
---

```
Package: rb3
Title: Download and Parse Public Data Released by B3 Exchange
Description: Download and parse public files released by B3 and convert them
    into useful formats and data structures common to data analysis
    practitioners.
Version: 0.0.3
Authors@R: c(person("Wilson", "Freitas",
                    email = "wilson.freitas@gmail.com",
                    role = c("aut", "cre")),
             person("Marcelo", "Perlin",
                    email = "marceloperlin@gmail.com",
                    role = "aut"))
    R (>= 4.1.0),
    bizdays,
    proto,
    cli,
    rvest,
    httr,
    ascii,
    yaml,
    digest,
    base64enc
    DT,
    miniUI,
    shiny,
    xtable,
    covr,
    scales,
    fixedincome
Collate:
    'rb3-package.R'
    'util.R'
    'transmute.R'
    'fields.R'
    'handlers.R'
    'marketdata.R'
    'download-data.R'
    'file.R'
    'convert_to.R'
    'scraper-cdi.R'
    'scraper-futures.R'
    'scraper-yc.R'
    'scraper-cotahist.R'
    'addin-show-templates.R'
    'addin-display-template.R'
    'readers.R'
    'downloaders.R'
    'zzz.R'
    http://wilsonfreitas.github.io/rb3/
```

The package downloads and organizes raw financial data directly from B3, the main financial exchange in Brazil. These datasets are not available in any other way.

Academic researchers and practioners of financial markets.

No.

Yes.
