### 2022-05-12 #534 @emilyriederer
Hello @msperlin and many thanks for your submission.

We are discussing whether the package is in scope and need a bit more information.

I see you are the author of other packages such as [GetDFDData2](https://cran.r-project.org/web/packages/GetDFPData2/index.html) already on CRAN. Could you please help to clarify the differences between these two packages? Is the former focused only on financial statements whereas this package provides more granular time series? While I'm sure your two packages are not "competing", I want to fully understand what makes this package novel and [how it is best in class](https://devguide.ropensci.org/policies.html#overlap).

Thank you!
