### 2022-05-12 #534 @msperlin
Hi @emilyriederer

Please find my replies below:

> I see you are the author of other packages such as [GetDFDData2](https://cran.r-project.org/web/packages/GetDFPData2/index.html) already on CRAN. Could you please help to clarify the differences between these two packages? Is the former focused only on financial statements whereas this package provides more granular time series? While I'm sure your two packages are not "competing", I want to fully understand what makes this package novel and [how it is best in class](https://devguide.ropensci.org/policies.html#overlap).

The different between GetDFPDAta2 and rb3 is in their scope and audience. Package GetDFPData2 is related to financial documents released by companies  to the exchange and its audience is mostly business analysts. This includes sales, profit, and many other financial information on a annual basis. Meanwhile, you cant think of rb3 as an interface to all **trade** data available at the b3 website, including prices of many different type of contracts such as options, equities (companies) and futures. ITs audience is more related to traders and market participants.

While there is some relationship between datasets from rb3 (stock prices) and GetDFPData2 (financial statements) for equities, there is none for other types of markets.

> Similarly, I see a B3 data listed on general services like [Yahoo Finance](https://finance.yahoo.com/quote/B3SA3.SA/) which I know can be accessed through a number of packages including `yfR`, `quantmod`, and `tidyquant`. Could you elaborate on why none of these packages are in a "comparison set"?

The data is indeed simillar, but only for equities (companies). Yahoo finance, for example, does not provide historical prices for futures or option contracts. Package rb3 also provides access to historical yield curves, which is a very unique (and rare) set of data in finance.
