### 2022-05-13 b8c5929
Added examples to functions show_templates and display_template
