+++ R/rb3-package.R
@@ -4,8 +4,6 @@
 #' Read the many files used in Brazilian Financial Market and
 #' convert them into useful formats and data structures.
 #'
-#' @name rb3
-#'
 #' @docType package
 #'
 #' @import proto
