### 2022-05-13 #534 @wilsonfreitas
Hi @emilyriederer and @msperlin

I added the examples to the functions show_templates and display_templates.
rb3 is not a function, it is just the tag `@name` being used in the package documentantion, I removed that.

And with respect to `<<-` operator, it is used with a ReferenceClass, and accoding to [documentation](https://www.rdocumentation.org/packages/methods/versions/3.6.2/topics/ReferenceClasses) this is the suggest way to modify objects fields.
So, I believe that this is not an issue.
