+++ R/rb3-package.R
@@ -4,6 +4,7 @@
 #' Read the many files used in Brazilian Financial Market and
 #' convert them into useful formats and data structures.
 #'
+#' @name rb3-package
 #' @docType package
 #'
 #' @import proto
