+++ man/rb3-package.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rb3-package.R
 \docType{package}
-\name{rb3}
-\alias{rb3}
+\name{rb3-package}
+\alias{rb3-package}
 \title{Read files from Brazilian Financial Market}
 \description{
 Read the many files used in Brazilian Financial Market and
