+++ tests/testthat/test-downloaders.R
@@ -1,5 +1,7 @@
 
-skip_on_cran()
+if (!covr::in_covr()) {
+  skip_on_cran()
+}
 
 test_that("it should download a file with a simple downloader", {
   tpl <- .retrieve_template(NULL, "CDIIDI")
@@ -18,4 +20,28 @@ test_that("it should download a file with a datetime downloader", {
   expect_true(file.exists(dest))
   info <- file.info(dest)
   expect_true(info$size > 1000000)
+})
+
+test_that("it should fail to datetime_download", {
+  tpl <- .retrieve_template(NULL, "OpcoesAcoesEmAberto")
+  f <- datetime_download(tpl, tempfile())
+  expect_false(f)
+})
+
+test_that("it should fail to settlement_prices_download", {
+  tpl <- .retrieve_template(NULL, "AjustesDiarios")
+  f <- settlement_prices_download(tpl, tempfile())
+  expect_false(f)
+})
+
+test_that("it should fail to curve_download", {
+  tpl <- .retrieve_template(NULL, "TaxasReferenciais")
+  f <- curve_download(tpl, tempfile())
+  expect_false(f)
+})
+
+test_that("it should defaults to PRE in curve_download", {
+  tpl <- .retrieve_template(NULL, "TaxasReferenciais")
+  f <- curve_download(tpl, tempfile(), refdate = as.Date("2022-05-10"))
+  expect_true(f)
 })
