### 2022-05-13 #534 @wilsonfreitas
Furtherly, the test coverage is now 84%.
