### 2022-05-13 65530e9
Added test for stock_indexes_composition_download
