### 2022-05-13 #534 @mpadge
Thanks @wilsonfreitas, our check system has just been updated to do what above link says: *Allow global assignment operator if in RefClass statement*. Your package no longer fails that check :+1:

And FYI, another new check has just been added which in your case reports this:

---

### 4. Other Checks

-  - `download_data` from admixr, FedData, fgeo.x, fHMM, hkdatasets, pRecipe, Rnumerai

---

That's currently a "trial" function so you don't necessarily need to respond, but feel free to change the name if you want.
