+++ legacy/formador-mercado.R
@@ -10,5 +10,5 @@ ch$HistoricalPrices |>
   filter(tipo_mercado %in% c(70, 80)) |>
   head()
 
-f <- download_data("COTAHIST_DAILY", refdate = Sys.Date() - 1)
+f <- download_marketdata("COTAHIST_DAILY", refdate = Sys.Date() - 1)
 ch <- read_marketdata(f, "COTAHIST_DAILY", parse_fields = TRUE)
