+++ R/download-data.R
@@ -20,13 +20,13 @@
 #'
 #' @examples
 #' \dontrun{
-#' fname <- download_data("CDIIDI")
+#' fname <- download_marketdata("CDIIDI")
 #' }
 #'
 #' @export
-download_data <- function(template,
-                          cache_folder = cachedir(),
-                          do_cache = TRUE, ...) {
+download_marketdata <- function(template,
+                                cache_folder = cachedir(),
+                                do_cache = TRUE, ...) {
   template <- .retrieve_template(NULL, template)
   x <- list(...)
   code_ <- digest::digest(x)
@@ -41,7 +41,7 @@ download_data <- function(template,
     return(fname)
   }
 
-  if (template$download_data(dest, ...)) {
+  if (template$download_marketdata(dest, ...)) {
     fname <- unzip_recursive(dest)
     return(fname)
   } else {
