### 2022-05-13 #534 @wilsonfreitas
Hi @mpadge

> Thanks @wilsonfreitas, our check system has just been updated to do what above link says: Allow global assignment operator if in RefClass statement. Your package no longer fails that check 👍

This is great!

---

* `download_data` has been renamed to `download_marketdata`.
