+++ README.md
@@ -21,7 +21,14 @@ status](https://www.r-pkg.org/badges/version/rb3)](https://CRAN.R-project.org/pa
 offering support and access to trading systems for equity and fixed
 income markets. In its website you can find a vast number of datasets
 regarding prices and transactions for contracts available for trading at
-these markets.
+these markets, including:
+
+-   equities (unadjusted for corporate events)
+-   futures
+-   FII (Reits)
+-   options
+-   BDRs
+-   historical yield curves (calculated from futures contracts)
 
 Package **rb3** facilitates downloading and reading these datasets from
 [B3](https://www.b3.com.br), making it easy to consume it in R in a
@@ -61,10 +68,6 @@ df_yc <- yc_mget(
   last_date = Sys.Date(),
   by = 255
 )
-#> Skipping download - using cached version
-#> Skipping download - using cached version
-#> Skipping download - using cached version
-#> Skipping download - using cached version
 
 p <- ggplot(
   df_yc,
@@ -112,14 +115,14 @@ glimpse(
 )
 #> Rows: 153
 #> Columns: 8
-#> $ refdate          <date> 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-~
-#> $ commodity        <chr> "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "~
-#> $ maturity_code    <chr> "J22", "K22", "M22", "N22", "Q22", "U22", "V22", "X22", "~
-#> $ symbol           <chr> "DI1J22", "DI1K22", "DI1M22", "DI1N22", "DI1Q22", "DI1U22~
-#> $ price_previous   <dbl> 99999.99, 99172.50, 98159.27, 97181.87, 96199.14, 95137.6~
-#> $ price            <dbl> 100000.00, 99172.31, 98160.23, 97185.43, 96210.42, 95159.~
-#> $ change           <dbl> 0.01, -0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 57.3~
-#> $ settlement_value <dbl> 0.01, 0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 57.32~
+#> $ refdate          <date> 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022…
+#> $ commodity        <chr> "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1…
+#> $ maturity_code    <chr> "J22", "K22", "M22", "N22", "Q22", "U22", "V22", "X22…
+#> $ symbol           <chr> "DI1J22", "DI1K22", "DI1M22", "DI1N22", "DI1Q22", "DI…
+#> $ price_previous   <dbl> 99999.99, 99172.50, 98159.27, 97181.87, 96199.14, 951…
+#> $ price            <dbl> 100000.00, 99172.31, 98160.23, 97185.43, 96210.42, 95…
+#> $ change           <dbl> 0.01, -0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, …
+#> $ settlement_value <dbl> 0.01, 0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 5…
 ```
 
 ### Equity data
@@ -130,6 +133,11 @@ thru `cotahist_get`.
 ``` r
 library(rb3)
 library(bizdays)
+#> 
[…]
