### 2022-05-14 97f3776
improved readme (fixed #34)
