+++ README.rmd
@@ -37,6 +37,7 @@ for contracts available for trading at these markets, including:
 
 Package **rb3** facilitates downloading and reading these datasets from 
 [B3](https://www.b3.com.br), making it easy to consume it in R in a structured way.
+These datasets can be used in industry or academic studies.
 
 
 # Documentation
