### 2022-05-14 #534 @msperlin
Thanks @emilyriederer. I added to the README.md the explicit datasets available at rb3. This should give context to the reader and set rb3 apart from other packages.
