### 2022-05-15 0b6b763
Organized and documented the template system
