+++ R/zzz.R
@@ -40,16 +40,20 @@ new_part <- function(x) {
 #' @importFrom utils getFromNamespace
 new_template <- function(tpl) {
   obj <- MarketData$proto()
+  obj[["has_reader"]] <- FALSE
+  obj[["has_downloader"]] <- FALSE
   for (n in names(tpl)) {
     if (n == "fields") {
       obj[["fields"]] <- do.call(fields, lapply(tpl$fields, new_field))
     } else if (n == "parts") {
       obj[["parts"]] <- lapply(tpl$parts, new_part)
     } else if (n == "reader") {
+      obj[["has_reader"]] <- TRUE
       obj[["reader"]] <- tpl$reader
       func_name <- tpl$reader[["function"]]
       obj[["read_file"]] <- getFromNamespace(func_name, "rb3")
     } else if (n == "downloader") {
+      obj[["has_downloader"]] <- TRUE
       obj[["downloader"]] <- tpl$downloader
       func_name <- tpl$downloader[["function"]]
       obj[["download_marketdata"]] <- getFromNamespace(func_name, "rb3")
@@ -61,6 +65,7 @@ new_template <- function(tpl) {
   if (is(try(obj$reader, TRUE), "try-error")) {
     reader_name <- paste0(stringr::str_to_lower(obj$filetype), "_read_file")
     obj[["read_file"]] <- getFromNamespace(reader_name, "rb3")
+    obj[["has_reader"]] <- TRUE
   }
 
   if (obj$filetype %in% c("MCSV", "MFWF")) {
