### 2022-05-15 5241574
New templates
- curves scenarios
- reference price scenarios
