+++ inst/extdata/templates/FPR.yaml
@@ -14,8 +14,6 @@ parts:
     fields:
     - name: tipo_registro
       description: Tipo de registro
-      handler:
-        type: character
     - name: data_geracao_arquivo
       description: Data de geração do arquivo
       handler:
@@ -26,16 +24,10 @@ parts:
     fields:
     - name: tipo_registro
       description: Tipo de registro
-      handler:
-        type: character
     - name: id_fpr
       description: ID do FPR
-      handler:
-        type: character
     - name: nome_fpr
       description: Nome do FPR
-      handler:
-        type: character
     - name: formato_variacao
       description: Formato de variação
       handler:
@@ -48,12 +40,8 @@ parts:
         - Basis Points
     - name: id_grupo_fpr
       description: ID do Grupo de FPR
-      handler:
-        type: character
     - name: id_camara_indicador
       description: Identificação da Câmara do Indicador
-      handler:
-        type: character
     - name: id_instrumento_indicador
       description: Identificação do Instrumento do Indicador
       handler:
@@ -62,17 +50,9 @@ parts:
         sign: ''
     - name: origem_instrumento
       description: Origem do Instrumento do Indicado
-      handler:
-        type: character
     - name: base
       description: Base
-      handler:
-        type: character
     - name: base_interpolacao
       description: Base de Interpolação
-      handler:
-        type: character
     - name: criterio_capitalizacao
       description: Critério de Capitalização
-      handler:
-        type: character
