+++ R/scraper-yc.R
@@ -2,7 +2,11 @@
 #'
 #' Downloads yield curve data from B3 website
 #' <https://www2.bmf.com.br/pages/portal/bmfbovespa/lumis/lum-taxas-referenciais-bmf-ptBR.asp>.
-#' Particularly, we import data for DI X Pre.
+#' Particularly, we import data for
+#' - DI X Pre (`yc_get`)
+#' - Cupom limpo (`yc_usd_get`)
+#' - DI x IPCA (`yc_ipca_get`)
+#' -
 #' See <https://www.b3.com.br/data/files/8B/F5/11/68/5391F61043E561F6AC094EA8/Manual_de_Curvas.pdf>
 #' for more details.
 #'
@@ -13,6 +17,7 @@
 #' @param cache_folder Location of cache folder (default = cachedir())
 #' @param do_cache Whether to use cache or not (default = TRUE)
 #'
+#' @details
 #' `yc_get` returns the yield curve for the given date and `yc_mget` returns
 #' multiple yield curves for a given range of dates.
 #'
@@ -106,8 +111,224 @@ get_single_yc <- function(idx_date,
           refdate,
           .data$biz_days, tpl$calendar
         ),
-        r_252 = df$r_252 / 100,
-        r_360 = df$r_360 / 100
+        r_252 = df$col1 / 100,
+        r_360 = df$col2 / 100
+      )
+    } else {
+      NULL
+    }
+  } else {
+    cli::cli_alert_danger("Error: no data found for date {refdate}")
+    return(NULL)
+  }
+}
+
+#' @details
+#' `yc_ipca_get` returns the yield curve of real interest rates
+#' for the given date and `yc_ipca_mget` returns
+#' multiple yield curves of real interest rates for a given range of dates.
+#' These real interest rates consider IPCA as its inflation index.
+#'
+#' @rdname yc_get
+#' @examples
+#' \dontrun{
+#' df_yc_ipca <- yc_ipca_mget(
+#'   first_date = Sys.Date() - 5,
+#'   last_date = Sys.Date()
+#' )
+#' head(df_yc_ipca)
+#' }
+#' @export
+yc_ipca_mget <- function(first_date = Sys.Date() - 5,
+                         last_date = Sys.Date(),
+                         by = 1,
+                         cache_folder = cachedir(),
+                         do_cache = TRUE) {
+  first_date <- as.Date(first_date)
+  last_date <- as.Date(last_date)
+
+  # find biz days in between
+  tpl <- .retrieve_template(NULL, "TaxasReferenciais")
+
+  date_vec <- bizdays::bizseq(first_date, last_date, tpl$calendar)
+
+  # use by to separate dates
+  date_vec <- date_vec[seq(1, length(date_vec), by = by)]
+
+  # get data!
+  df_yc <- dplyr::bind_rows(
+    purrr::map(cli::cli_progress_along(
+      date_vec,
+      format = paste0(
+        "{pb_spin} Fetching data points ",
+        " {cli::pb_current}/{cli::pb_total} ",
+        " | {pb_bar} {pb_percent} | {pb_eta_str}"
+      )
+    ),
+    get_single_yc_ipca,
+    date_vec = date_vec,
+    cache_folder = cache_folder,
+    do_cache = do_cache
+    )
+  )
+
+  return(df_yc)
+}
+
+#' @rdname yc_get
+#' @examples
+#' \dontrun{
+#' df_yc_ipca <- yc_ipca_get(Sys.Date())
+#' head(df_yc_ipca)
+#' }
+#' @export
+yc_ipca_get <- function(refdate = Sys.Date(),
+                        cache_folder = cachedir(),
+                        do_cache = TRUE) {
+  get_single_yc_ipca(1, as.Date(refdate), cache_folder, do_cache)
+}
+
[…]
