+++ tests/testthat/test-yc.R
@@ -57,3 +57,102 @@ test_that("Test of yc_get function", {
   expect_identical(df_yc_1, df_yc_2)
 })
 
+test_that("Test of yc_ipca_mget function", {
+  if (!covr::in_covr()) {
+    skip_on_cran()
+    skip_if_offline()
+  }
+
+  first_date <- Sys.Date() - 10
+  last_date <- Sys.Date()
+
+  # first call (no cache)
+  df_yc_1 <- yc_ipca_mget(first_date,
+    last_date,
+    by = 2,
+    do_cache = FALSE
+  )
+
+  test_df(df_yc_1)
+
+  # first call (with cache)
+  df_yc_2 <- yc_ipca_mget(first_date,
+    last_date,
+    by = 2
+  )
+
+  test_df(df_yc_2)
+
+  expect_identical(df_yc_1, df_yc_2)
+})
+
+test_that("Test of yc_ipca_get function", {
+  if (!covr::in_covr()) {
+    skip_on_cran()
+    skip_if_offline()
+  }
+
+  refdate <- bizdays::offset(Sys.Date(), -30, "Brazil/ANBIMA")
+
+  # first call (no cache)
+  df_yc_1 <- yc_ipca_get(refdate, do_cache = FALSE)
+
+  test_df(df_yc_1)
+
+  # first call (with cache)
+  df_yc_2 <- yc_ipca_get(refdate)
+
+  test_df(df_yc_2)
+
+  expect_identical(df_yc_1, df_yc_2)
+})
+
+test_that("Test of yc_usd_mget function", {
+  if (!covr::in_covr()) {
+    skip_on_cran()
+    skip_if_offline()
+  }
+
+  first_date <- Sys.Date() - 10
+  last_date <- Sys.Date()
+
+  # first call (no cache)
+  df_yc_1 <- yc_usd_mget(first_date,
+    last_date,
+    by = 2,
+    do_cache = FALSE
+  )
+
+  test_df(df_yc_1)
+
+  # first call (with cache)
+  df_yc_2 <- yc_usd_mget(first_date,
+    last_date,
+    by = 2
+  )
+
+  test_df(df_yc_2)
+
+  expect_identical(df_yc_1, df_yc_2)
+})
+
+test_that("Test of yc_usd_get function", {
+  if (!covr::in_covr()) {
+    skip_on_cran()
+    skip_if_offline()
+  }
+
+  refdate <- bizdays::offset(Sys.Date(), -30, "Brazil/ANBIMA")
+
+  # first call (no cache)
+  df_yc_1 <- yc_usd_get(refdate, do_cache = FALSE)
+
+  test_df(df_yc_1)
+
+  # first call (with cache)
+  df_yc_2 <- yc_usd_get(refdate)
+
+  test_df(df_yc_2)
+
+  expect_identical(df_yc_1, df_yc_2)
+})
