+++ NAMESPACE
@@ -24,7 +24,11 @@ export(maturity2date)
 export(read_marketdata)
 export(show_templates)
 export(yc_get)
+export(yc_ipca_get)
+export(yc_ipca_mget)
 export(yc_mget)
+export(yc_usd_get)
+export(yc_usd_mget)
 import(bizdays)
 import(cli)
 import(dplyr)

[…]
