+++ README.md
@@ -29,6 +29,7 @@ these markets, including:
 -   options
 -   BDRs
 -   historical yield curves (calculated from futures contracts)
+-   B3 indexes composition
 
 Package **rb3** facilitates downloading and reading these datasets from
 [B3](https://www.b3.com.br), making it easy to consume it in R in a
@@ -69,10 +70,6 @@ df_yc <- yc_mget(
   last_date = Sys.Date(),
   by = 255
 )
-#> Skipping download - using cached version
-#> Skipping download - using cached version
-#> Skipping download - using cached version
-#> Skipping download - using cached version
 
 p <- ggplot(
   df_yc,
@@ -120,14 +117,14 @@ glimpse(
 )
 #> Rows: 153
 #> Columns: 8
-#> $ refdate          [3m[38;5;246m<date>[39m[23m 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-0~
-#> $ commodity        [3m[38;5;246m<chr>[39m[23m "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1~
-#> $ maturity_code    [3m[38;5;246m<chr>[39m[23m "J22", "K22", "M22", "N22", "Q22", "U22", "V22", "X22", "Z22", "F23", "G23", "H23", "J23", "N23", "V23", "F24", "J24", "N24", "V24~
-#> $ symbol           [3m[38;5;246m<chr>[39m[23m "DI1J22", "DI1K22", "DI1M22", "DI1N22", "DI1Q22", "DI1U22", "DI1V22", "DI1X22", "DI1Z22", "DI1F23", "DI1G23", "DI1H23", "DI1J23", ~
-#> $ price_previous   [3m[38;5;246m<dbl>[39m[23m 99999.99, 99172.50, 98159.27, 97181.87, 96199.14, 95137.64, 94174.49, 93265.23, 92365.48, 91404.64, 90434.90, 89662.14, 88719.94, ~
-#> $ price            [3m[38;5;246m<dbl>[39m[23m 100000.00, 99172.31, 98160.23, 97185.43, 96210.42, 95159.25, 94209.42, 93314.08, 92422.80, 91472.04, 90513.73, 89751.01, 88821.78,~
-#> $ change           [3m[38;5;246m<dbl>[39m[23m 0.01, -0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 57.32, 67.40, 78.83, 88.87, 101.84, 150.68, 175.83, 275.55, 370.18, 423.00, 4~
-#> $ settlement_value [3m[38;5;246m<dbl>[39m[23m 0.01, 0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 57.32, 67.40, 78.83, 88.87, 101.84, 150.68, 175.83, 275.55, 370.18, 423.00, 45~
+#> $ refdate          <date> 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022-04~
[…]
