### 2022-05-16 3fedd8e
New index functions
- index_comp_get
- indexes_get
- indexes_last_update
- cotahist_get_symbols - to get stocks by symbols

Tests and docs are ok.
