+++ man/indexes_get.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/scraper-indexes.R
+\name{indexes_get}
+\alias{indexes_get}
+\title{Get B3 indexes available}
+\usage{
+indexes_get(cache_folder = cachedir(), do_cache = TRUE)
+}
+\arguments{
+\item{cache_folder}{Location of cache folder (default = cachedir())}
+
+\item{do_cache}{Whether to use cache or not (default = TRUE)}
+}
+\value{
+a character vector with symbols of indexes available
+}
+\description{
+Gets B3 indexes available.
+}
+\examples{
+indexes_get()
+}
