+++ NAMESPACE
@@ -11,6 +11,7 @@ export(cotahist_equity_options_get)
 export(cotahist_funds_get)
 export(cotahist_funds_options_get)
 export(cotahist_get)
+export(cotahist_get_symbols)
 export(cotahist_index_options_get)
 export(cotahist_indexes_get)
 export(cotahist_units_get)
@@ -20,6 +21,9 @@ export(download_marketdata)
 export(futures_get)
 export(futures_mget)
 export(idi_get)
+export(index_comp_get)
+export(indexes_get)
+export(indexes_last_update)
 export(maturity2date)
 export(read_marketdata)
 export(show_templates)
