### 2022-05-18 966c221
Changed path construction
Avoid string concatenation, using vector instead.
