+++ R/downloaders.R
@@ -69,7 +69,8 @@ stock_indexes_composition_download <- function(., dest, ...) {
   ), auto_unbox = TRUE)
   params_enc <- base64enc::base64encode(charToRaw(params))
   url <- httr::parse_url(.$downloader$url)
-  url$path <- paste0(url$path, "/", params_enc)
+  url$path <- c(url$path, params_enc)
+  res <- httr::GET(url)
   res <- httr::GET(url)
   if (httr::status_code(res) != 200) {
     return(FALSE)
