### 2022-05-18 0c6fc84
New template GetTheoricalPortfolio
It gets data with index compositon and its weigths.

Added functions:
- stock_indexes_theo_portfolio_download
- stock_indexes_json_reader

Changes:
- created new filetype MCUSTOM for multipart in custom files (zzz.R)
