+++ inst/extdata/templates/GetTheoricalPortfolio.yaml
@@ -1,0 +1,37 @@
+id: GetTheoricalPortfolio
+filename: ~
+filetype: MCUSTOM
+description: Carteira Teórica dos índices da B3
+locale: pt
+downloader:
+  function: stock_indexes_theo_portfolio_download
+  url: https://sistemaswebb3-listados.b3.com.br/indexProxy/indexCall/GetTheoricalPortfolio
+  format: json
+  encoding: latin1
+reader:
+  function: stock_indexes_json_reader
+parts:
+  Header:
+    name: header
+    fields:
+    - name: part
+      description: Participação % no índice
+    - name: theoricalQty
+      description: Quantidade teórica no índice
+    - name: reductor
+      description: Redutor
+  Results:
+    name: results
+    fields:
+    - name: code
+      description: Código da ação
+    - name: company
+      description: Nome da companhia
+    - name: spotlight
+      description: Código de especificação da ação
+    - name: theoricalQty
+      description: Quantidade teórica no índice
+    - name: part
+      description: Participação % no índice
+    - name: cont
+      description: Contagem
