### 2022-05-18 9803daa
GetStockIndex is now MCUSTOM
It uses reader stock_indexes_json_reader that is generic for indexes JSONs.
