### 2022-05-18 925892e
Added readers tests for GetTheoricalPortfolio
