### 2022-05-18 68ea854
New function index_weights_get
