+++ man/index_weights_get.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/scraper-indexes.R
-\name{index_comp_get}
-\alias{index_comp_get}
-\title{Get composition of B3 indexes}
+\name{index_weights_get}
+\alias{index_weights_get}
+\title{Get the assets weights of B3 indexes}
 \usage{
-index_comp_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
+index_weights_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
 }
 \arguments{
 \item{index_name}{a string with the index name}
@@ -14,13 +14,14 @@ index_comp_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
 \item{do_cache}{Whether to use cache or not (default = TRUE)}
 }
 \value{
-a character vector with symbols that belong to the given index name
+data.frame with symbols that belong to the given index name with its weights
+and theorical positions.
 }
 \description{
-Gets the composition of listed B3 indexes.
+Gets the assets weights of B3 indexes.
 }
 \examples{
 \dontrun{
-index_comp_get("IBOV")
+index_weights_get("IBOV")
 }
 }
