+++ NAMESPACE
@@ -22,6 +22,7 @@ export(futures_get)
 export(futures_mget)
 export(idi_get)
 export(index_comp_get)
+export(index_weights_get)
 export(indexes_get)
 export(indexes_last_update)
 export(maturity2date)
@@ -57,6 +58,7 @@ importFrom(rvest,html_table)
 importFrom(rvest,html_text)
 importFrom(rvest,read_html)
 importFrom(utils,getFromNamespace)
+importFrom(utils,hasName)
 importFrom(utils,read.table)
 importFrom(utils,unzip)
 importFrom(utils,write.table)
