### 2022-05-18 6e57a19
New template GetPortfolioDay
