+++ R/downloaders.R
@@ -103,3 +103,32 @@ stock_indexes_theo_portfolio_download <- function(., dest, ...) {
   save_resource(res, enc, dest)
   TRUE
 }
+
+stock_indexes_current_portfolio_download <- function(., dest, ...) {
+  args <- list(...)
+  if (!hasName(args, "index_name")) {
+    cli::cli_alert_danger("index_name argument not provided")
+    return(FALSE)
+  }
+  segment <- 2
+  if (hasName(args, "segment")) {
+    segment <- args$segment
+  }
+  params <- jsonlite::toJSON(list(
+    pageNumber = 1,
+    pageSize = 9999,
+    language = "pt-br",
+    index = args$index_name,
+    segment = segment
+  ), auto_unbox = TRUE)
+  params_enc <- base64encode(charToRaw(params))
+  url <- httr::parse_url(.$downloader$url)
+  url$path <- c(url$path, params_enc)
+  res <- httr::GET(url)
+  if (httr::status_code(res) != 200) {
+    return(FALSE)
+  }
+  enc <- if (is.null(.$downloader$encoding)) "utf8" else .$downloader$encoding
+  save_resource(res, enc, dest)
+  TRUE
+}
