### 2022-05-18 bd1fb48
Template: GetPortfolioDay - change encoding to UTF-8
