### 2022-05-18 d572c9a
New function index_by_segment_get
