### 2022-05-19 95cd795
New functions
- cotahist_fiagros_get
- cotahist_fidcs_get
- cotahist_fiis_get
- cotahist_etfs_get

The function cotahist_funds_get has been removed.
