+++ NAMESPACE
@@ -8,7 +8,10 @@ export(convert_to)
 export(cotahist_bdrs_get)
 export(cotahist_equity_get)
 export(cotahist_equity_options_get)
-export(cotahist_funds_get)
+export(cotahist_etfs_get)
+export(cotahist_fiagros_get)
+export(cotahist_fidcs_get)
+export(cotahist_fiis_get)
 export(cotahist_funds_options_get)
 export(cotahist_get)
 export(cotahist_get_symbols)
