+++ README.md
@@ -70,6 +70,10 @@ df_yc <- yc_mget(
   last_date = Sys.Date(),
   by = 255
 )
+#> Skipping download - using cached version
+#> Skipping download - using cached version
+#> Skipping download - using cached version
+#> Skipping download - using cached version
 
 p <- ggplot(
   df_yc,
@@ -135,11 +139,6 @@ thru `cotahist_get`.
 ``` r
 library(rb3)
 library(bizdays)
-#> 
-#> Attaching package: 'bizdays'
-#> The following object is masked from 'package:stats':
-#> 
-#>     offset
 
 # fix for ssl error (only in linux)
 if (Sys.info()["sysname"] == "Linux") {
@@ -150,48 +149,72 @@ if (Sys.info()["sysname"] == "Linux") {
 
 date <- preceding(Sys.Date() - 1, "Brazil/ANBIMA") # last business day
 ch <- cotahist_get(date, "daily")
+#> Skipping download - using cached version
 
 glimpse(
   cotahist_equity_get(ch)
 )
-#> Rows: 382
+#> Rows: 381
 #> Columns: 13
-#> $ refdate               <date> 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 2022-05-13, 20~
-#> $ symbol                <chr> "AALR3", "ABCB4", "ABEV3", "AERI3", "AESB3", "AFLT3", "AGRO3", "AGXY3", "ALLD3", "ALPA3", "ALPA4", "ALPK3", "ALSO3", "ALUP3", "ALUP4", "AMAR3", "AM~
-#> $ open                  <dbl> 19.81, 16.16, 14.55, 3.83, 10.79, 9.06, 31.60, 9.23, 13.35, 19.52, 22.28, 3.79, 19.65, 8.81, 8.66, 2.17, 28.52, 22.93, 5.02, 29.29, 12.36, 80.70, 1~
-#> $ high                  <dbl> 19.89, 16.55, 14.67, 3.93, 10.81, 9.50, 32.40, 9.78, 13.35, 19.52, 22.46, 3.80, 19.90, 8.90, 8.94, 2.28, 29.98, 24.62, 5.47, 32.98, 13.24, 84.32, 1~
-#> $ low                   <dbl> 19.43, 16.06, 14.45, 3.72, 10.64, 9.06, 31.54, 9.21, 12.30, 18.83, 21.70, 3.65, 19.48, 8.52, 8.60, 2.17, 28.52, 22.74, 5.01, 29.29, 12.24, 80.59, 1~
-#> $ close                 <dbl> 19.75, 16.54, 14.54, 3.75, 10.64, 9.35, 32.06, 9.33, 12.52, 19.25, 21.80, 3.80, 19.69, 8.70, 8.75, 2.23, 29.75, 23.05, 5.38, 32.98, 13.00, 82.49, 1~
-#> $ average               <dbl> 19.66, 16.38, 14.55, 3.80, 10.69, 9.38, 31.99, 9.50, 12.73, 19.12, 22.04, 3.75, 19.66, 8.72, 8.76, 2.23, 29.51, 23.52, 5.35, 31.46, 13.02, 82.75, 1~
[…]
