+++ NAMESPACE
@@ -24,6 +24,7 @@ export(download_marketdata)
 export(futures_get)
 export(futures_mget)
 export(idi_get)
+export(index_by_segment_get)
 export(index_comp_get)
 export(index_weights_get)
 export(indexes_get)
