### 2022-05-19 2cbbf30
Missing .data to dplyr call
