+++ tests/testthat/test-indexes.R
@@ -7,8 +7,8 @@ test_that("it should get index weights", {
   x <- index_weights_get("IBOV")
   expect_s3_class(x, "data.frame")
   expect_true(ncol(x) == 3)
-  expect_equal(colnames(x), c("symbols", "weights", "position"))
-  expect_equal(as.integer(sum(x$weights)), 1L)
+  expect_equal(colnames(x), c("symbol", "weight", "position"))
+  expect_equal(as.integer(sum(x$weight)), 1L)
   expect_true(nrow(x) > 0)
 })
 
