### 2022-05-20 f4ce5a5
Changed cotahist_equity_options_get
cotahist_equity_options_get also returns options whose underlyings are units.

cotahist_units_options_get has been removed
