+++ NAMESPACE
@@ -18,7 +18,6 @@ export(cotahist_get_symbols)
 export(cotahist_index_options_get)
 export(cotahist_indexes_get)
 export(cotahist_units_get)
-export(cotahist_units_options_get)
 export(display_template)
 export(download_marketdata)
 export(futures_get)
