### 2022-05-20 e0b3ab0
Updated tests to improve coverage
