### 2022-05-20 efde438
Implemented cotahist_equity_options_superset
Issue #30
