+++ NAMESPACE
@@ -8,6 +8,7 @@ export(convert_to)
 export(cotahist_bdrs_get)
 export(cotahist_equity_get)
 export(cotahist_equity_options_get)
+export(cotahist_equity_options_superset)
 export(cotahist_etfs_get)
 export(cotahist_fiagros_get)
 export(cotahist_fidcs_get)
