+++ NAMESPACE
@@ -35,9 +35,12 @@ export(show_templates)
 export(yc_get)
 export(yc_ipca_get)
 export(yc_ipca_mget)
+export(yc_ipca_superset)
 export(yc_mget)
+export(yc_superset)
 export(yc_usd_get)
 export(yc_usd_mget)
+export(yc_usd_superset)
 import(bizdays)
 import(cli)
 import(dplyr)
