### 2022-05-21 46ca07b
New template IndexReport
