+++ NAMESPACE
@@ -50,6 +50,9 @@ import(purrr)
 import(readr)
 import(stringr)
 import(yaml)
+importFrom(XML,getNodeSet)
+importFrom(XML,xmlInternalTreeParse)
+importFrom(XML,xmlValue)
 importFrom(base64enc,base64encode)
 importFrom(digest,digest)
 importFrom(dplyr,tibble)
