### 2022-05-22 be55a12
Added calendar and locale to IndexReport template
