### 2022-05-22 d0e78f6
New functions indexreport_get and indexreport_mget
