+++ R/util.R
@@ -28,3 +28,32 @@ clearcache <- function() {
   cache_folder <- cachedir()
   unlink(cache_folder, recursive = TRUE)
 }
+
+#' Fetches a single marketdata
+#'
+#' @param idx_date index of data (1.. n_dates)
+#' @param date_vec Vector of dates
+#' @param cache_folder Location of cache folder (default = cachedir())
+#' @param do_cache Whether to use cache or not (default = TRUE)
+#' @param ... orther arguments
+#'
+#' @return
+#' A dataframe or `NULL`
+#'
+#' @noRd
+get_single_marketdata <- function(template,
+                                  idx_date,
+                                  date_vec,
+                                  cache_folder,
+                                  do_cache, ...) {
+  refdate <- date_vec[idx_date]
+  fname <- download_marketdata(template, cache_folder, do_cache,
+    refdate = refdate, ...
+  )
+  if (!is.null(fname)) {
+    read_marketdata(fname, template, TRUE, cache_folder, do_cache)
+  } else {
+    cli::cli_alert_danger("Error: no data found for date {refdate}")
+    return(NULL)
+  }
+}
