+++ tests/testthat/test-indexes.R
@@ -1,3 +1,8 @@
+if (!covr::in_covr()) {
+  skip_on_cran()
+  skip_if_offline()
+}
+
 test_that("it should get index composition", {
   x <- index_comp_get("IBOV")
   expect_true(length(x) > 0)
@@ -27,5 +32,18 @@ test_that("it should get index by segments", {
   expect_s3_class(x, "data.frame")
   expect_true(ncol(x) == 6)
   expect_equal(as.integer(sum(x$weight)), 1L)
+  expect_true(nrow(x) > 0)
+})
+
+test_that("it should get indexreport", {
+  date <- preceding(Sys.Date() - 1, "Brazil/B3")
+  x <- suppressWarnings(indexreport_get(date))
+  expect_s3_class(x, "data.frame")
+  expect_true(ncol(x) == 8)
+  expect_true(nrow(x) > 0)
+  date1 <- preceding(Sys.Date() - 5, "Brazil/B3")
+  x <- suppressWarnings(indexreport_mget(date1, date))
+  expect_s3_class(x, "data.frame")
+  expect_true(ncol(x) == 8)
   expect_true(nrow(x) > 0)
 })
