+++ NAMESPACE
@@ -29,6 +29,8 @@ export(index_comp_get)
 export(index_weights_get)
 export(indexes_get)
 export(indexes_last_update)
+export(indexreport_get)
+export(indexreport_mget)
 export(maturity2date)
 export(read_marketdata)
 export(show_templates)
