### 2022-05-25 6cf5e0d
New template PriceReport
