### 2022-05-25 c67dd5f
Changed transmuter regex to coerce int
To convert numbers to integers the number can't be greater that 8 characters.
