### 2022-05-27 e31b0b6
New template GetListedSupplementCompany
new functions to organize the code structure:
- check_parameters
- url_encoded_download

These functions can be used to download with the urls that use base64 encoding.
