+++ R/downloaders.R
@@ -105,27 +105,63 @@ stock_indexes_theo_portfolio_download <- function(., dest, ...) {
 }
 
 stock_indexes_current_portfolio_download <- function(., dest, ...) {
-  args <- list(...)
-  if (!hasName(args, "index_name")) {
-    cli::cli_alert_danger("index_name argument not provided")
+  if (!check_parameters(..., arg_name = "index_name")) {
     return(FALSE)
   }
+  args <- list(...)
   segment <- 2
   if (hasName(args, "segment")) {
     segment <- args$segment
   }
-  params <- jsonlite::toJSON(list(
+  url_encoded_download(., dest,
     pageNumber = 1,
     pageSize = 9999,
     language = "pt-br",
     index = args$index_name,
     segment = segment
-  ), auto_unbox = TRUE)
+  )
+}
+
+company_listed_supplement_download <- function(., dest, ...) {
+  if (!check_parameters(..., arg_name = "company_name")) {
+    return(FALSE)
+  }
+  args <- list(...)
+  url_encoded_download(., dest,
+    issuingCompany = args$company_name, language = "pt-br"
+  )
+}
+
+company_details_download <- function(., dest, ...) {
+  if (!check_parameters(..., arg_name = "code_cvm")) {
+    return(FALSE)
+  }
+  args <- list(...)
+  url_encoded_download(., dest,
+    codeCVM = args$code_cvm, language = "pt-br"
+  )
+}
+
+#' @importFrom stringr str_glue
+check_parameters <- function(..., arg_name) {
+  args <- list(...)
+  if (!hasName(args, arg_name)) {
+    msg <- str_glue("{arg_name} argument not provided")
+    cli::cli_alert_danger(msg)
+    FALSE
+  } else {
+    TRUE
+  }
+}
+
+url_encoded_download <- function(., dest, ...) {
+  args <- list(...)
+  params <- toJSON(args, auto_unbox = TRUE)
   params_enc <- base64encode(charToRaw(params))
-  url <- httr::parse_url(.$downloader$url)
+  url <- parse_url(.$downloader$url)
   url$path <- c(url$path, params_enc)
-  res <- httr::GET(url)
-  if (httr::status_code(res) != 200) {
+  res <- GET(url)
+  if (status_code(res) != 200) {
     return(FALSE)
   }
   enc <- if (is.null(.$downloader$encoding)) "utf8" else .$downloader$encoding
