### 2022-05-27 f2fbb48
Organized stock_* functions
- stock_indexes_composition_download
- stock_indexes_theo_portfolio_download
- stock_indexes_current_portfolio_download

These functions have been organized to reduce repetitive code.
