+++ R/downloaders.R
@@ -63,45 +63,24 @@ curve_download <- function(., dest, ...) {
 }
 
 stock_indexes_composition_download <- function(., dest, ...) {
-  params <- jsonlite::toJSON(list(
+  url_encoded_download(., dest,
     pageNumber = 1,
     pageSize = 9999
-  ), auto_unbox = TRUE)
-  params_enc <- base64enc::base64encode(charToRaw(params))
-  url <- httr::parse_url(.$downloader$url)
-  url$path <- c(url$path, params_enc)
-  res <- httr::GET(url)
-  if (httr::status_code(res) != 200) {
-    return(FALSE)
-  }
-  enc <- if (is.null(.$downloader$encoding)) "utf8" else .$downloader$encoding
-  save_resource(res, enc, dest)
-  TRUE
+  )
 }
 
 #' @importFrom utils hasName
 stock_indexes_theo_portfolio_download <- function(., dest, ...) {
-  args <- list(...)
-  if (!hasName(args, "index_name")) {
-    cli::cli_alert_danger("index_name argument not provided")
+  if (!check_parameters(..., arg_name = "index_name")) {
     return(FALSE)
   }
-  params <- jsonlite::toJSON(list(
+  args <- list(...)
+  url_encoded_download(., dest,
     pageNumber = 1,
     pageSize = 9999,
     language = "pt-br",
     index = args$index_name
-  ), auto_unbox = TRUE)
-  params_enc <- base64enc::base64encode(charToRaw(params))
-  url <- httr::parse_url(.$downloader$url)
-  url$path <- c(url$path, params_enc)
-  res <- httr::GET(url)
-  if (httr::status_code(res) != 200) {
-    return(FALSE)
-  }
-  enc <- if (is.null(.$downloader$encoding)) "utf8" else .$downloader$encoding
-  save_resource(res, enc, dest)
-  TRUE
+  )
 }
 
 stock_indexes_current_portfolio_download <- function(., dest, ...) {
