+++ legacy/fatos-relevantes.R
@@ -49,7 +49,7 @@ base64enc::base64decode(hsh) |>
 
 res <- paste0(
   "https://sistemaswebb3-listados.b3.com.br/listedCompaniesProxy/CompanyCall/GetListedSupplementCompany/",
-  list(issuingCompany = "VALE", language = "pt-br") |>
+  list(issuingCompany = "BBAS", language = "pt-br") |>
     jsonlite::toJSON(auto_unbox = TRUE) |>
     charToRaw() |>
     base64enc::base64encode()
@@ -60,16 +60,25 @@ res
 (x <- httr::content(res, as = "text") |> jsonlite::fromJSON())
 x$cashDividends[[1]]
 x$stockDividends[[1]]
+x$subscriptions[[1]]
 x$stockCapital
 x$segment
 x$quotedPerSharSince
+x$commonSharesForm
+x$preferredSharesForm
+x$hasCommom
+x$hasPreferred
 x$code
 x$codeCVM
 x$totalNumberShares
 x$numberCommonShares
 x$numberPreferredShares
-x$subscriptions
+x$roundLot
+x$tradingName
+names(x)
 
+(y <- httr::content(res, as = "text") |> jsonlite::fromJSON(simplifyDataFrame = FALSE))
+View(y)
 
 # ----
 
@@ -81,6 +90,20 @@ base64enc::base64decode(hsh) |>
   rawToChar() |>
   jsonlite::fromJSON()
 
+res <- paste0(
+  "https://sistemaswebb3-listados.b3.com.br/listedCompaniesProxy/CompanyCall/GetDetail/",
+  list(codeCVM = x$codeCVM, language = "pt-br") |>
+    jsonlite::toJSON(auto_unbox = TRUE) |>
+    charToRaw() |>
+    base64enc::base64encode()
+) |> httr::GET(httr::verbose())
+
+res
+
+(g <- httr::content(res, as = "text") |> jsonlite::fromJSON())
+names(g)
+View(g)
+
 # https://sistemaswebb3-listados.b3.com.br/listedCompaniesProxy/CompanyCall/GetListedFinancial/eyJjb2RlQ1ZNIjoiNDE3MCIsImxhbmd1YWdlIjoicHQtYnIifQ==
 
 hsh <- "eyJjb2RlQ1ZNIjoiNDE3MCIsImxhbmd1YWdlIjoicHQtYnIifQ=="
@@ -88,3 +111,57 @@ hsh <- "eyJjb2RlQ1ZNIjoiNDE3MCIsImxhbmd1YWdlIjoicHQtYnIifQ=="
 base64enc::base64decode(hsh) |>
   rawToChar() |>
   jsonlite::fromJSON()
+
+res <- paste0(
+  "https://sistemaswebb3-listados.b3.com.br/listedCompaniesProxy/CompanyCall/GetListedFinancial/",
+  list(codeCVM = x$codeCVM, language = "pt-br") |>
+    jsonlite::toJSON(auto_unbox = TRUE) |>
+    charToRaw() |>
+    base64enc::base64encode()
+) |> httr::GET(httr::verbose())
+
+res
+
+(f <- httr::content(res, as = "text") |> jsonlite::fromJSON())
+names(f)
+f$titleInitial
+f$consolidated |> as_tibble()
+f$unconsolidated |> as_tibble()
+f$freeFloatResult
+
[…]
