### 2022-05-29 421a83f
New template GetListedCashDividends
