### 2022-05-29 34fe977
company_cash_dividends_download argument tradingName
The argument tradingName can have only: space, digits and capital letters
Other characters must be replaced with "".
