+++ NAMESPACE
@@ -69,6 +69,7 @@ importFrom(rvest,html_nodes)
 importFrom(rvest,html_table)
 importFrom(rvest,html_text)
 importFrom(rvest,read_html)
+importFrom(stringr,str_glue)
 importFrom(utils,getFromNamespace)
 importFrom(utils,hasName)
 importFrom(utils,read.table)
