### 2022-05-29 645a835
added locale `en` to templates: `COTAHIST_*`
the column `traded_contracts` has more than 8 digits so it must be coerced to numeric, instead of integer.

These templates didnt have locale declared so they didn't had a generic rule to convert big numbers.

The locale specific rules do that.
