+++ inst/extdata/templates/COTAHIST_YEARLY.yaml
@@ -2,6 +2,7 @@ id: COTAHIST_YEARLY
 filename: COTAHIST
 filetype: MFWF
 description: Cotações Históricas do Pregão de Ações - Arquivo Anual
+locale: en
 downloader:
   function: datetime_download
   url: https://bvmf.bmfbovespa.com.br/InstDados/SerHist/COTAHIST_A%Y.ZIP
