+++ inst/extdata/templates/COTAHIST_MONTHLY.yaml
@@ -2,6 +2,7 @@ id: COTAHIST_MONTHLY
 filename: COTAHIST
 filetype: MFWF
 description: Cotações Históricas do Pregão de Ações - Arquivo Mensal
+locale: en
 downloader:
   function: datetime_download
   url: https://bvmf.bmfbovespa.com.br/InstDados/SerHist/COTAHIST_M%m%Y.ZIP
