### 2022-06-05 a7fe2c2
New Template NegociosIntraday
To create this template I have created a new handler for strings representing time with up to nine digits.

The handler type is strtime and one new function was created to parse the column.
