### 2022-06-05 38f95a6
Adding cli namespace
