+++ R/scraper-indexes.R
@@ -180,9 +180,9 @@ indexreport_mget <- function(first_date = Sys.Date() - 5,
     map(cli::cli_progress_along(
       date_vec,
       format = paste0(
-        "{pb_spin} Fetching data points ",
-        " {cli::pb_current}/{cli::pb_total} ",
-        " | {pb_bar} {pb_percent} | {pb_eta_str}"
+        "{cli::pb_spin} Fetching data points",
+        "{cli::pb_current}/{cli::pb_total}",
+        " | {cli::pb_bar} {cli::pb_percent} | {cli::pb_eta_str}"
       )
     ),
     get_single_indexreport,
