+++ tests/testthat/test-downloaders.R
@@ -51,3 +51,41 @@ test_that("it should defaults to PRE in curve_download", {
   f <- curve_download(tpl, tempfile(), refdate = as.Date("2022-05-10"))
   expect_true(f)
 })
+
+test_that("it should base64_datetime_download", {
+  tpl <- .retrieve_template(NULL, "NegociosBalcao")
+  refdate <- preceding(Sys.Date() - 1, "Brazil/B3")
+  f <- base64_datetime_download(tpl, tempfile(), refdate = refdate)
+  expect_true(f)
+})
+
+test_that("it should fail base64_datetime_download", {
+  tpl <- .retrieve_template(NULL, "NegociosBalcao")
+  refdate <- as.Date("2022-06-05")
+  f <- base64_datetime_download(tpl, tempfile(), refdate = refdate)
+  expect_false(f)
+})
+
+test_that("it should company_listed_supplement_download", {
+  tpl <- .retrieve_template(NULL, "GetListedSupplementCompany")
+  f <- company_listed_supplement_download(tpl, tempfile(),
+    company_name = "ABEV"
+  )
+  expect_true(f)
+})
+
+test_that("it should company_details_download", {
+  tpl <- .retrieve_template(NULL, "GetDetailsCompany")
+  f <- company_details_download(tpl, tempfile(),
+    code_cvm = "23264"
+  )
+  expect_true(f)
+})
+
+test_that("it should company_cash_dividends_download ", {
+  tpl <- .retrieve_template(NULL, "GetListedCashDividends")
+  f <- company_cash_dividends_download(tpl, tempfile(),
+    trading_name = "AMBEVSA"
+  )
+  expect_true(f)
+})
