### 2022-06-07 9f61ffc
Added function to NAMESPACE
