+++ NAMESPACE
@@ -91,6 +91,7 @@ importFrom(rvest,html_nodes)
 importFrom(rvest,html_table)
 importFrom(rvest,html_text)
 importFrom(rvest,read_html)
+importFrom(stringr,str_c)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_ends)
 importFrom(stringr,str_glue)
