### 2022-06-07 4e51ba4
Added check for linux on test-downloaders
