### 2022-06-17 34bba85
Added option rb3.cachedir
In order to keep cached files accross sessions, the option rb3.cachedir has been created.
