### 2022-06-17 977828b
Updated NEWS
