+++ NEWS.md
@@ -6,6 +6,7 @@
   * `NegociosBalcao` intraday OTC (Debentures) trades
   * `NegociosBTB` intraday lending trades
 * imports organized (using importFrom in NAMESPACE)
+* added option `rb3.cachedir` to set default cache directory in order to use cached files accross multiple sessions
 
 
 # rb3 0.0.3 
