### 2022-06-20 ac07611
changed to handle empty dividends
