+++ examples/companies-info.R
@@ -4,27 +4,44 @@ library(bizdays)
 library(purrr)
 
 ch <- cotahist_get(preceding(Sys.Date() - 1, "Brazil/ANBIMA"), "daily")
-
 eqs <- cotahist_equity_get(ch)
 
-eqs |> distinct(symbol)
+parse_isin <- function(isin) {
+  tibble(
+    isin = isin,
+    country = str_sub(isin, 1, 2),
+    asset_name = str_sub(isin, 3, 6),
+    asset_type = str_sub(isin, 7, 9),
+    spec_type = str_sub(isin, 10, 11),
+    control = str_sub(isin, 12)
+  ) |> mutate(
+    spec_type = str_replace(spec_type, "PR", "PN"),
+    spec_type = str_replace(spec_type, "OR", "ON")
+  )
+}
 
-company_info_get <- function(symbol) {
-  cat(symbol, "\n")
+create_codes <- function(codes) {
+  left_join(codes, parse_isin(codes$isin), "isin") |>
+    rename(symbol = code)
+}
+
+.company_supplement_get <- function(code) {
   template <- "GetListedSupplementCompany"
-  f <- download_marketdata(template, company_name = symbol)
-  company_info <- try(read_marketdata(f, template), TRUE)
-  if (is(company_info, "try-error")) {
-    f <- download_marketdata(template, company_name = str_sub(symbol, 1, 4))
-    company_info <- try(read_marketdata(f, template), TRUE)
-  }
+  f <- download_marketdata(template, company_name = code)
+  read_marketdata(f, template)
+}
+
+.company_info_get <- function(code) {
+  company_info <- .company_supplement_get(code)
+
   template <- "GetDetailsCompany"
   f <- download_marketdata(template, code_cvm = company_info$Info$codeCVM)
   company_details <- read_marketdata(f, template)
 
-  segment <- str_split(company_details$Info$industryClassification, "/")
-  segment <- list(segment[[1]] |> str_trim())
-  df <- tibble(
+  sectors <- str_split(company_details$Info$industryClassification, "/")
+  sectors <- sectors[[1]] |> str_trim()
+  tibble(
+    asset_name = company_info$Info$code,
     trading_name = company_info$Info$tradingName,
     company_name = company_details$Info$companyName,
     activity = company_details$Info$activity,
@@ -33,42 +50,259 @@ company_info_get <- function(symbol) {
     total_shares = company_info$Info$totalNumberShares,
     common_shares = company_info$Info$numberCommonShares,
     preferred_shares = company_info$Info$numberPreferredShares,
-    codes = list(company_details$OtherCodes$code),
-    segment = segment,
+    codes = list(company_details$OtherCodes),
+    sector = sectors[1],
+    subsector = sectors[2],
+    market_segment = sectors[3],
+    round_lot = company_info$Info$roundLot,
+    quoted_since = company_info$Info$quotedPerShareSince,
+    segment = company_info$Info$segment
   )
-  df$id <- digest::digest(df)
-  df$symbol <- symbol
-  df
 }
 
-company_df <- map(eqs$symbol, company_info_get)
+company_info_get <- function(symbols) {
+  codes <- tibble(
+    symbol = symbols,
+    asset_name = str_sub(symbol, 1, 4)
+  ) |>
+    group_by(asset_name) |>
+    summarise(symbols = list(symbol))
 
-company_df <- do.call(rbind, company_df)
+  rxx <- function(x, idx = 0) {
+    codes_ <- codes[x, ]
+    res <- if (idx == 0) {
+      try(.company_info_get(codes_$asset_name), TRUE)
+    } else {
[…]
