### 2022-06-22 5194bdc
GetListedSupplementCompany updated
Part key subscription renamed to Subscription with capital S
