### 2022-06-22 50341ff
test updated
BIDI11 is in process to leave IBOV and at 2022-06-21 it had no trade, but it is still in the IBOV composition and that test was leading to a failure.
