+++ examples/companies-info.R
@@ -1,29 +1,22 @@
-library(rb3)
-library(tidyverse)
-library(bizdays)
-library(purrr)
 
-ch <- cotahist_get(preceding(Sys.Date() - 1, "Brazil/ANBIMA"), "daily")
-eqs <- cotahist_equity_get(ch)
+cotahist_equity_symbols_get <- function(ch) {
+  df <- ch[["HistoricalPrices"]] |>
+    filter(
+      .data$tipo_mercado %in% 10,
+      str_sub(.data$cod_isin, 7, 9) %in% c("UNT", "CDA", "ACN")
+    )
 
-df <- ch[["HistoricalPrices"]] |>
-  filter(
-    .data$tipo_mercado %in% 10,
-    str_sub(.data$cod_isin, 7, 9) %in% c("UNT", "CDA", "ACN")
-  )
-
-spec_type <- df[["especificacao"]] |>
-  str_split("\\s+") |>
-  map_chr(\(x) x[1])
-
-market_segment <- df[["especificacao"]] |>
-  str_split("\\s+") |>
-  map_chr(\(x) x[length(x)])
-
-codes <- parse_isin(df[["cod_isin"]])
-codes[["spec_type2"]] <- spec_type
-
-
+  spec_split <- df[["especificacao"]] |> str_split("\\s+")
+  codes <- parse_isin(df[["cod_isin"]])
+  codes[["spec_type"]] <- spec_split |> map_chr(\(x) x[1])
+  codes[["symbol"]] <- df[["cod_negociacao"]]
+  codes |>
+    select(
+      symbol, asset_name, spec_type, isin_spec_type, isin, country
+    ) |>
+    unique() |>
+    arrange(symbol)
+}
 
 smartget <- function(key, dict) {
   x <- try(get(key, dict, inherits = FALSE), TRUE)
@@ -40,11 +33,8 @@ parse_isin <- function(isin) {
     country = str_sub(isin, 1, 2),
     asset_name = str_sub(isin, 3, 6),
     asset_type = str_sub(isin, 7, 9),
-    spec_type = str_sub(isin, 10, 11),
+    isin_spec_type = str_sub(isin, 10, 11),
     control = str_sub(isin, 12)
-  ) |> mutate(
-    spec_type = str_replace(spec_type, "PR", "PN"),
-    spec_type = str_replace(spec_type, "OR", "ON")
   )
 }
 
@@ -125,15 +115,13 @@ company_info_get <- function(symbols) {
   }
 
   company_info$StockDividends |>
-    left_join(company_details$OtherCodes, by = c("isinCode" = "isin")) |>
     rename(
-      symbol = code,
       isin = isinCode,
       approved = approvedOn,
       last_date_prior_ex = lastDatePrior,
       description = label
     ) |>
-    select(symbol, isin, approved, last_date_prior_ex, description, factor)
+    select(isin, approved, last_date_prior_ex, description, factor)
 }
 
 company_stock_dividends_get <- function(symbols) {
@@ -173,9 +161,7 @@ company_stock_dividends_get <- function(symbols) {
   }
 
   company_info$Subscriptions |>
-    left_join(company_details$OtherCodes, by = c("isinCode" = "isin")) |>
     rename(
-      symbol = code,
       isin = isinCode,
       approved = approvedOn,
       last_date_prior_ex = lastDatePrior,
@@ -185,7 +171,7 @@ company_stock_dividends_get <- function(symbols) {
       subscription_date = subscriptionDate,
     ) |>
     select(
-      symbol, isin, approved, last_date_prior_ex, description, percentage,
+      isin, approved, last_date_prior_ex, description, percentage,
       trading_period, price_unit, subscription_date
     )
 }
[…]
