### 2022-07-02 11cf259
Corrected bug in wrong cache use
Issue #44
