+++ examples/companies-info.R
@@ -79,28 +79,7 @@ create_codes <- function(codes) {
 }
 
 company_info_get <- function(symbols) {
-  codes <- tibble(
-    symbol = symbols,
-    asset_name = str_sub(symbol, 1, 4)
-  ) |>
-    group_by(asset_name) |>
-    summarise(symbols = list(symbol))
-
-  rxx <- function(x, idx = 0) {
-    codes_ <- codes[x, ]
-    res <- if (idx == 0) {
-      try(.company_info_get(codes_$asset_name), TRUE)
-    } else {
-      try(.company_info_get(codes_$symbols[[1]][idx]), TRUE)
-    }
-    if (is(res, "try-error") && length(codes_$symbols[[1]]) >= idx + 1) {
-      rxx(x, idx + 1)
-    } else {
-      res
-    }
-  }
-  companies_list <- map(seq_len(nrow(codes)), rxx)
-  bind_rows(companies_list)
+  .company_mget(.company_info_get, symbols)
 }
 
 .company_stock_dividends_get <- function(code) {
@@ -125,28 +104,7 @@ company_info_get <- function(symbols) {
 }
 
 company_stock_dividends_get <- function(symbols) {
-  codes <- tibble(
-    symbol = symbols,
-    asset_name = str_sub(symbol, 1, 4)
-  ) |>
-    group_by(asset_name) |>
-    summarise(symbols = list(symbol))
-
-  rxx <- function(x, idx = 0) {
-    codes_ <- codes[x, ]
-    res <- if (idx == 0) {
-      try(.company_stock_dividends_get(codes_$asset_name), TRUE)
-    } else {
-      try(.company_stock_dividends_get(codes_$symbols[[1]][idx]), TRUE)
-    }
-    if (is(res, "try-error") && length(codes_$symbols[[1]]) >= idx + 1) {
-      rxx(x, idx + 1)
-    } else {
-      res
-    }
-  }
-  companies_list <- map(seq_len(nrow(codes)), rxx)
-  bind_rows(companies_list)
+  .company_mget(.company_stock_dividends_get, symbols)
 }
 
 .company_subscriptions_get <- function(code) {
@@ -177,36 +135,34 @@ company_stock_dividends_get <- function(symbols) {
 }
 
 company_subscriptions_get <- function(symbols) {
-  codes <- tibble(
-    symbol = symbols,
-    asset_name = str_sub(symbol, 1, 4)
-  ) |>
-    group_by(asset_name) |>
-    summarise(symbols = list(symbol))
+  .company_mget(.company_subscriptions_get, symbols)
+}
 
-  rxx <- function(x, idx = 0) {
-    codes_ <- codes[x, ]
-    res <- if (idx == 0) {
-      try(.company_subscriptions_get(codes_$asset_name), TRUE)
-    } else {
-      try(.company_subscriptions_get(codes_$symbols[[1]][idx]), TRUE)
-    }
-    if (is(res, "try-error") && length(codes_$symbols[[1]]) >= idx + 1) {
-      rxx(x, idx + 1)
-    } else {
-      res
-    }
+.company_info_cash_dividends_get <- function(code) {
+  company_info <- .company_supplement_get(code)
+
+  # Proventos distribuídos pelo emissor nos últimos 12 meses ou o último,
+  # se anterior aos 12 últimos meses.
+  if (!is.null(company_info$CashDividends)) {
+    company_info$CashDividends |>
+      mutate(ratio = 1) |>
+      rename(
+        isin = isinCode,
+        approved = approvedOn,
+        last_date_prior_ex = lastDatePrior,
+        description = label,
+        value_cash = rate,
+      ) |>
+      select(
[…]
