+++ man/clearcache.Rd
@@ -13,7 +13,7 @@ Has no return
 Clear cache directory
 }
 \examples{
-\dontrun {
+\dontrun{
 clearcache()
 }
 }
