### 2022-07-16 729e859
BUG due to change in fixedincome
fixedincome function rates has been changed to implied_rate.
