+++ vignettes/Fetching-historical-future-rates.Rmd
@@ -58,7 +58,7 @@ di1_futures <- df |>
         maturity_date = maturity2date(maturity_code),
         fixing = following(maturity_date, "Brazil/ANBIMA"),
         business_days = bizdays(refdate, maturity_date, "Brazil/ANBIMA"),
-        adjusted_tax = rates("discrete", business_days / 252, 100000 / price)
+        adjusted_tax = implied_rate("discrete", business_days / 252, 100000 / price)
     ) |>
     filter(business_days > 0)
 
@@ -90,7 +90,7 @@ dap_futures <- df |>
         maturity_date = maturity2date(maturity_code, "15th day"),
         fixing = following(maturity_date, "Brazil/ANBIMA"),
         business_days = bizdays(refdate, maturity_date, "Brazil/ANBIMA"),
-        adjusted_tax = rates("discrete", business_days / 252, 100000 / price)
+        adjusted_tax = implied_rate("discrete", business_days / 252, 100000 / price)
     ) |>
     filter(business_days > 0)
 
@@ -129,7 +129,7 @@ infl_futures <- df |>
         ),
         fixing = following(maturity_date, "Brazil/ANBIMA"),
         business_days = bizdays(refdate, maturity_date, "Brazil/ANBIMA"),
-        adjusted_tax = rates("discrete", business_days / 252, 100000 / price)
+        adjusted_tax = implied_rate("discrete", business_days / 252, 100000 / price)
     ) |>
     arrange(refdate)
 
