### 2022-07-16 ea26008
Salvando arquivos por template no cache
Para implementar isso foi necessário mudar a assinatura de read_marketdata removendo o argumento cache_folder.
O arquivo gerado é salvo no mesmo diretório do arquivo fornecido.
O argumento do_cache funciona da mesma forma verificando se o arquivo já existe e caso exista ele é utilizado e o arquivo fornecido não é parsed.

Issue #46
