+++ R/scraper-futures.R
@@ -91,7 +91,7 @@ futures_mget <- function(first_date = Sys.Date() - 5,
     map(cli::cli_progress_along(
       date_vec,
       format = paste0(
-        "{cli::pb_spin} Fetching data points",
+        "{cli::pb_spin} Fetching data points ",
         "{cli::pb_current}/{cli::pb_total}",
         " | {cli::pb_bar} {cli::pb_percent} | {cli::pb_eta_str}"
       )
@@ -126,7 +126,7 @@ single_futures_get <- function(idx_date,
   refdate <- date_vec[idx_date]
   fname <- download_marketdata(tpl, cache_folder, do_cache, refdate = refdate)
   if (!is.null(fname)) {
-    df <- read_marketdata(fname, tpl, TRUE, cache_folder, do_cache)
+    df <- read_marketdata(fname, tpl, TRUE, do_cache)
     if (!is.null(df)) {
       tibble(
         refdate = as.Date(refdate),
