+++ tests/testthat/test-bd_arbit.R
@@ -1,6 +1,7 @@
 
 test_that("it should load BD_Arbit with the correct field types", {
   f_txt <- system.file("extdata/BD_Arbit.txt", package = "rb3")
+  f_txt <- copy_file_to_temp(f_txt)
 
   df <- read_marketdata(f_txt, template = "BD_Arbit")
   expect_s3_class(df$data_geracao_arquivo, "Date")
