+++ tests/testthat/test-indic.R
@@ -1,6 +1,7 @@
 
 test_that("read file by template name", {
   f <- system.file("extdata/Indic.txt", package = "rb3")
+  f <- copy_file_to_temp(f)
   res <- read_marketdata(f, template = "Indic")
 
   classes <- c(
