+++ NEWS.md
@@ -1,3 +1,7 @@
+# rb3 0.0.6
+
+* Updated documentation
+
 # rb3 0.0.5
 
 * Updated documentation
