### 2022-07-17 2a381c5
Set rb3.cachedir to tests
Also organized clearcache tests
