+++ tests/testthat/test-cache.R
@@ -1,11 +1,8 @@
 test_that("cache fcts", {
-  
-  this_cachedir <- 	cachedir()
-  expect_true(class(this_cachedir) == 'character')
-  
-  clearcache() 
-  
-  # if passed clearcache(), set true
-  expect_true(TRUE)
-  
+  this_cachedir <- cachedir()
+  expect_true(class(this_cachedir) == "character")
+  expect_true(dir.exists(this_cachedir))
+
+  clearcache()
+  expect_false(dir.exists(this_cachedir))
 })
