### 2022-07-17 c0bee85
Functions code2month and maturity2date accept old codes
BMF had codes for months with characters of length 3.
The current codes have length 1.

The functions have been addapted to handle both codes.
