### 2022-07-17 ebaef99
functions `code2month` and `maturity2date` now accept old 4 characters maturity code, before 2006
