### 2022-07-18 65b96d2
Documentation updated
Updated documentation for functions code2month and maturity2date.

Added str_length to NAMESPACE
