+++ NAMESPACE
@@ -95,6 +95,7 @@ importFrom(stringr,str_c)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_ends)
 importFrom(stringr,str_glue)
+importFrom(stringr,str_length)
 importFrom(stringr,str_match)
 importFrom(stringr,str_pad)
 importFrom(stringr,str_replace)
