### 2022-07-18 f30385b
Added test fixure
Created setup.R file on tests.

Using withr to implement a test fixure.

Added withr to Suggests.
