+++ _pkgdown.yml
@@ -1,4 +1,55 @@
 url: http://wilsonfreitas.github.io/rb3/
 template:
-  bootstrap: 5
+  params:
+    bootswatch: yeti
+
+reference:
+- title: rb3 Package
+  contents:
+  - rb3-package
+
+- title: Download and Read Marketdata
+  contents:
+  - read_marketdata
+  - download_marketdata
+  - convert_to
+
+- title: Templates
+  contents:
+  - show_templates
+  - display_template
+
+- title: COTAHIST datasets
+  contents:
+  - cotahist_get
+  - cotahist_equity_get
+
+- title: Indexes
+  contents:
+  - indexes_get
+  - indexes_comp_get
+  - indexes_weights_get
+  - indexes_by_segment_get
+  - indexes_last_update
+  - indexreport_get
+
+- title: Futures
+  contents:
+  - futures_get
+  - maturity2date
+  - code2month
+
+- title: CDI rate and IDI index
+  contents:
+  - cdi_get
+
+- title: Super datasets (mix datasets)
+  contents:
+  - yc_superset
+  - cotahist_equity_options_superset
+
+- title: Cache
+  contents:
+  - cachedir
+  - clearcache
 
