+++ _pkgdown.yml
@@ -27,10 +27,10 @@ reference:
 - title: Indexes
   contents:
   - indexes_get
-  - indexes_comp_get
-  - indexes_weights_get
-  - indexes_by_segment_get
   - indexes_last_update
+  - index_comp_get
+  - index_weights_get
+  - index_by_segment_get
   - indexreport_get
 
 - title: Futures
