### 2022-08-13 0da0c3a
Fix images in documentation
Issue #28
