### 2022-08-13 bb3ca3e
Don't cache NULL data
Issue #52
