### 2022-08-13 6f60556
cdi_get caches downloaded data, but it shouldn't
Issue #51
