### 2022-08-13 d3bf5c8
Created cotahist_options_by_symbol_superset
This function returns a superset with all options accordingly to the passed symbol.

It works with equities and ETF.

Issued #50
