+++ NAMESPACE
@@ -18,6 +18,7 @@ export(cotahist_get)
 export(cotahist_get_symbols)
 export(cotahist_index_options_get)
 export(cotahist_indexes_get)
+export(cotahist_options_by_symbol_superset)
 export(cotahist_units_get)
 export(display_template)
 export(download_marketdata)
