+++ DESCRIPTION
@@ -73,6 +73,6 @@ URL: https://github.com/wilsonfreitas/rb3,
     http://wilsonfreitas.github.io/rb3/
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.1
 Config/testthat/edition: 3
 Encoding: UTF-8
