### 2022-08-13 2b53949
Updated NEWS
