### 2022-08-16 2f34e07
Organized covr usage
Removed the check for covr in tests and set environment variable NOT_CRAN to not skip the tests when running covr.
