+++ tests/testthat/test-addin.R
@@ -1,10 +1,0 @@
-test_that("rstudio addin", {
-  
-  # make sure the functions are called in covr::report()
-  if (covr::in_covr()) {
-    show_templates()
-    display_template()
-  }
-  
-  expect_true(TRUE)
-})
