+++ tests/testthat/test-downloaders.R
@@ -1,7 +1,6 @@
 
-if (!covr::in_covr()) {
-  skip_on_cran()
-}
+skip_on_cran()
+skip_if_offline()
 
 if (Sys.info()["sysname"] == "Linux") {
   httr::set_config(httr::config(ssl_verifypeer = FALSE))
