### 2022-08-28 98ac80a
Index historical data
Issue #39

- new template GetPortfolioDay_IndexStatistics to download indexes historical data
- IBOVDIA.XLS IBOVESPA historical data from 1968-1997
- IBOV.rds dataframe with historical data from 1968-1997
- functions to access index historical data
- documentation
