+++ R/rb3-package.R
@@ -14,11 +14,13 @@
 #' @importFrom bizdays add.bizdays bizdayse bizseq getdate
 #' @importFrom digest digest
 #' @importFrom dplyr tibble inner_join mutate select filter left_join as_tibble
-#' @importFrom dplyr bind_rows
+#' @importFrom dplyr bind_rows arrange
+#' @importFrom tidyr pivot_longer
 #' @importFrom httr GET POST parse_url status_code headers content
 #' @importFrom jsonlite toJSON fromJSON
 #' @importFrom purrr map_dfr map_lgl map_chr map_int map
 #' @importFrom readr write_rds read_rds read_csv read_file
+#' @importFrom readxl read_excel
 #' @importFrom rlang .data
 #' @importFrom stringr str_replace_all str_starts str_match str_sub str_split
 #' @importFrom stringr str_to_lower str_detect str_pad str_replace str_trim
