### 2022-08-28 655f6a2
Added index_get to pkgdown
