### 2022-08-28 c8149e1
New vignette B3 Indexes
