### 2022-08-29 #534 @mpadge
@msperlin and @emilyriederer you may ignore the "functions do not document return values" fail, as that has only come online long after this package was submitted. @msperlin you may of course update your code to document these, as they'll also fail CRAN checks for new packages, but that's up to you. The duplicated names checks may also be ingored for the moment.
