### 2022-08-29 #534 @emilyriederer
Thanks for the clarification, @mpadge ! With that, I think it sounds like we should be ready to proceed with the review. I will begin to work towards assigning a handling editor.
