### 2022-09-06 #534 @emilyriederer
Hi @msperlin and @wilsonfreitas - since my EiC term has ended, I have self-volunteered to be the editor for this package as well. I will begin considering reviewers, but in the mean time, please let me direct your attention to a few items found in the automated checks:
