### 2022-09-06 #534 @emilyriederer
---

I think this package is quite close to being read for review. I just want to note a few core issues:

- Since this is an API-heavy package, please consider the [rOpenSci guidance](https://books.ropensci.org/http-testing/) for http testing best practices. I think some of the tools listed there could be helpful to make the current checks more robust
- In the `CONTRIBUTING.md`, I notice there are no instructions about setting up API keys for testing. Can you please confirm that users do not need to provide any authentication information for these APIs?

Additionally, from the automated checks, I want to highlight a few items in particular:

- Some stated imports (e.g. `cli`) are not found as being used. Can this dependency be removed?
- Please review and resolve the formatting issues found by `lintr`
- You may consider whether it is a concern that `index_get()` is a function name used in other packages. I doubt if this package would be used much with `elastic` so I don't foresee the conflict arising frequently, but if another name suits you equally well, it might be better to change

---
