### 2022-09-08 #534 @wilsonfreitas
Hi @emilyriederer,

Thanks for your suggestions.
I'm answering your questions here.

> Tests: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)?
Ok, I see valuable points here, I'll work on them.

> Project management: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled?
Yes, it is, but we don't have an structured process to be managed.

I'm going to work on the other points: dependencies, formatting, function names ...
Thanks
