### 2022-09-20 664749e
Resolved formating issues found by lintr
