+++ vignettes/Fetching-historical-future-rates.Rmd
@@ -153,7 +153,8 @@ ranges from Jan. 2023 to Jan. 2033.
 df_fut <- df |>
     filter(symbol %in% c("DI1F23", "DI1F33")) |>
     mutate(
-        maturity_date = maturity2date(maturity_code) |> following("Brazil/ANBIMA"),
+        maturity_date = maturity2date(maturity_code) |>
+            following("Brazil/ANBIMA"),
         business_days = bizdays(refdate, maturity_date, "Brazil/ANBIMA")
     )
 
