+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   B3-Indexes: B3-Indexes.html
   Fetching-historical-future-rates: Fetching-historical-future-rates.html
   Fetching-historical-yield-curve: Fetching-historical-yield-curve.html
-last_built: 2022-08-28T21:52Z
+last_built: 2022-09-20T11:40Z
 urls:
   reference: http://wilsonfreitas.github.io/rb3/reference
   article: http://wilsonfreitas.github.io/rb3/articles
