### 2022-09-21 #534 @emilyriederer
Hi @wilsonfreitas ! I hope you are well. I thought I'd follow up and see if you had a chance to make any updates in response to my comments? Most all of my comments were a bit open-ended and not a hard, binary requirement. Once you've incorporated them to your satisfaction, I can begin to look for reviewers.
