### 2022-09-21 #534 @wilsonfreitas
Hi @emilyriederer, thanks!!

Unfortunately I am slower than I'd like, but I have a few updates.

> In the CONTRIBUTING.md, I notice there are no instructions about setting up API keys for testing. Can you please confirm that users do not need to provide any authentication information for these APIs?

The package makes heavy use of web scraping, it does not require any API to access the data and no setup is necessary for testing.
Should I make this clear in the CONTRIBUTING.md?

> Please review and resolve the formatting issues found by lintr

Done!

> You may consider whether it is a concern that index_get() is a function name used in other packages. I doubt if this package would be used much with elastic so I don't foresee the conflict arising frequently, but if another name suits you equally well, it might be better to change

If this is not blocking I'd prefer to keep `index_get` as function name.

> Some stated imports (e.g. cli) are not found as being used. Can this dependency be removed?

Currently, I am working on this point.

> Since this is an API-heavy package, please consider the [rOpenSci guidance](https://books.ropensci.org/http-testing/) for http testing best practices. I think some of the tools listed there could be helpful to make the current checks more robust

I am reading this book to see what can be applied in the package.
