### 2022-09-23 f4c6007
Using direct arguments in str_glue
Using arguments directly in str_glue to avoid lintr notificactions pointing unused variables.
