### 2022-09-24 4428dcb
Organized alert system to avoid unnecessary alert calls
cli::cli calls has been wrapped into functions to control the alert display with options rb3.silent and rb3.hide.progressbar.
