+++ R/util.R
@@ -68,7 +68,78 @@ get_single_marketdata <- function(template,
   if (!is.null(fname)) {
     read_marketdata(fname, template, TRUE, do_cache)
   } else {
-    cli::cli_alert_danger("Error: no data found for date {refdate}")
+    alert("danger", "Error: no data found for date {refdate}",
+      refdate = refdate
+    )
     return(NULL)
+  }
+}
+
+#' cli_progress_along wrapper
+#'
+#' @param x data to iterate through
+#' @param func function to call
+#' @param msg message to display
+#' @param ... orther arguments
+#'
+#' @return
+#' A list with `func` returned values
+#'
+#' @noRd
+log_map_process_along <- function(x, func, msg, ...) {
+  f_ <- paste(
+    "{pb_spin}",
+    "{msg}",
+    "{pb_current}/{pb_total}",
+    "|",
+    "{pb_bar}",
+    "{pb_percent}",
+    "|",
+    "{pb_eta_str}"
+  )
+  rb3_hide_progressbar <- getOption("rb3.hide.progressbar")
+  if (!is.null(rb3_hide_progressbar) && isTRUE(rb3_hide_progressbar)) {
+    map(seq_along(x), func, ...)
+  } else {
+    map(cli_progress_along(x, format = f_), func, ...)
+  }
+}
+
+#' cli_alert_* wrapper
+#'
+#' @param x type
+#' @param text message to display
+#' @param ... orther arguments
+#'
+#' @return
+#' A list with `func` returned values
+#'
+#' @noRd
+alert <- function(x = c("info", "success", "danger", "warning"), text, ...) {
+  x <- match.arg(x)
+  rb3_silent <- getOption("rb3.silent")
+  if (!is.null(rb3_silent) && isTRUE(rb3_silent)) {
+    # do nothing
+  } else {
+    f_ <- alert_fun(x)
+    if (! is.null(f_)) {
+      f_(str_glue(text, .envir = list(...)))
+    }
+  }
+}
+
+alert_fun <- function(x) {
+  funcs <- list(
+    info = cli_alert_info,
+    danger = cli_alert_danger,
+    success = cli_alert_success,
+    warning = cli_alert_warning
+  )
+  func <- funcs[[x]]
+  if (is.null(func)) {
+    warning(paste0("Invalid call to alert function ", x))
+    return(NULL)
+  } else {
+    func
   }
 }
