+++ R/scraper-indexes.R
@@ -181,18 +181,11 @@ indexreport_mget <- function(first_date = Sys.Date() - 5,
   date_vec <- date_vec[seq(1, length(date_vec), by = by)]
 
   bind_rows(
-    map(cli::cli_progress_along(
-      date_vec,
-      format = paste0(
-        "{cli::pb_spin} Fetching data points",
-        "{cli::pb_current}/{cli::pb_total}",
-        " | {cli::pb_bar} {cli::pb_percent} | {cli::pb_eta_str}"
-      )
-    ),
-    get_single_indexreport,
-    date_vec = date_vec,
-    cache_folder = cache_folder,
-    do_cache = do_cache
+    log_map_process_along(date_vec, get_single_indexreport,
+      "Fetching data points",
+      date_vec = date_vec,
+      cache_folder = cache_folder,
+      do_cache = do_cache
     )
   )
 }
