### 2022-09-25 85d0307
Let's keep rb3.silent as the only option to control message display
