+++ tests/testthat/setup.R
@@ -4,7 +4,6 @@ withr::defer(unlink(local_cachedir, recursive = TRUE), teardown_env())
 
 op <- options(
   rb3.cachedir = local_cachedir,
-  rb3.hide.progressbar = TRUE,
   rb3.silent = TRUE
 )
 withr::defer(options(op), teardown_env())
