+++ R/rb3-package.R
@@ -4,6 +4,33 @@
 #' Read the many files used in Brazilian Financial Market and
 #' convert them into useful formats and data structures.
 #'
+#' @details
+#' ## rb3 options
+#'
+#' rb3 uses `base::options` to allow user set global options that affect the
+#' way the package works and display its alerts.
+#'
+#' \describe{
+#'   \item{rb3.cachedir}{
+#'     rb3 cache folder is named `rb3-cache` and it is created inside the
+#'     directory returned by `base::tempdir`.
+#'     Since it is changed for every new session it is interesting to use the
+#'     same directory for cache accross sessions.
+#'     Once the option `rb3.cachedir` is set the files are always cached in
+#'     the same directory.
+#'     This is very useful to build a historical data.
+#'     Historical time series can be loaded directly from cached files.
+#'   }
+#'   \item{rb3.silent}{
+#'     rb3 default behavior on communicating users what's going on is total
+#'     transparency.
+#'     So, it displays many alert messages to inform users many of the details.
+#'     On the other hand, this behavior can be sometimes annoying.
+#'     The option `rb3.silent` can be set to `TRUE` in order to avoid that the
+#'     alerts be displayed.
+#'   }
+#' }
+#'
 #' @name rb3-package
 #' @docType package
 #'
