### 2022-09-25 #534 @wilsonfreitas
Hi @emilyriederer

I think I am done for now.
The last commits have many changes implementing your suggestions.
We can go forward with the process.
