+++ README.md
@@ -15,6 +15,8 @@ coverage](https://codecov.io/gh/wilsonfreitas/rb3/branch/main/graph/badge.svg)](
 [![CRAN
 status](https://www.r-pkg.org/badges/version/rb3)](https://CRAN.R-project.org/package=rb3)
 [![](https://cranlogs.r-pkg.org/badges/rb3)](https://cran.r-project.org/package=rb3)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/534_status.svg)](https://github.com/ropensci/software-review/issues/534)
 <!-- badges: end -->
 
 [B3](https://www.b3.com.br) is the main financial exchange in Brazil,
@@ -126,14 +128,14 @@ glimpse(
 )
 #> Rows: 153
 #> Columns: 8
-#> $ refdate          <date> 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022…
-#> $ commodity        <chr> "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1…
-#> $ maturity_code    <chr> "J22", "K22", "M22", "N22", "Q22", "U22", "V22", "X22…
-#> $ symbol           <chr> "DI1J22", "DI1K22", "DI1M22", "DI1N22", "DI1Q22", "DI…
-#> $ price_previous   <dbl> 99999.99, 99172.50, 98159.27, 97181.87, 96199.14, 951…
-#> $ price            <dbl> 100000.00, 99172.31, 98160.23, 97185.43, 96210.42, 95…
-#> $ change           <dbl> 0.01, -0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, …
-#> $ settlement_value <dbl> 0.01, 0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 5…
+#> $ refdate          <date> 2022-04-01, 2022-04-01, 2022-04-01, 2022-04-01, 2022~
+#> $ commodity        <chr> "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1", "DI1~
+#> $ maturity_code    <chr> "J22", "K22", "M22", "N22", "Q22", "U22", "V22", "X22~
+#> $ symbol           <chr> "DI1J22", "DI1K22", "DI1M22", "DI1N22", "DI1Q22", "DI~
+#> $ price_previous   <dbl> 99999.99, 99172.50, 98159.27, 97181.87, 96199.14, 951~
+#> $ price            <dbl> 100000.00, 99172.31, 98160.23, 97185.43, 96210.42, 95~
+#> $ change           <dbl> 0.01, -0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, ~
+#> $ settlement_value <dbl> 0.01, 0.19, 0.96, 3.56, 11.28, 21.61, 34.93, 48.85, 5~
 ```
 
 ### Equity data
@@ -163,21 +165,21 @@ ch <- cotahist_get(date, "daily")
 glimpse(
   cotahist_equity_get(ch)
 )
-#> Rows: 383
+#> Rows: 391
 #> Columns: 13
-#> $ refdate               <date> 2022-07-25, 2022-07-25, 2022-07-25, 2022-07-25,…
-#> $ symbol                <chr> "AALR3", "ABCB4", "ABEV3", "AERI3", "AESB3", "AG…
-#> $ open                  <dbl> 19.88, 16.42, 14.53, 2.97, 10.30, 23.43, 7.10, 9…
-#> $ high                  <dbl> 20.18, 16.92, 14.78, 2.97, 10.50, 23.74, 7.22, 1…
[…]
