### 2022-10-03 #534 @emilyriederer
@wilsonfreitas , please meet your reviewers @quishqa and @pachadotdev . I am particularly excited to have them reviewing this package as they are both rOpenSci package developers themselves with packages that support different data extraction tasks focused on South America. Lots of commonalities and opportunities to share best practices here!
