### 2022-10-19 #534 @pachadotdev
hi!

just an initial round of questions

1. can you please add vcr to suggests? I would prefer to be able to test offline without skipping the test but having the cassettes to run the test anyways, as in https://github.com/ropensci/tradestatistics/blob/master/tests/testthat/test-ots_create_tidy_data.R#L9

2. do you have a roadmap to test the API-centric functions? these seem to be the most important but here only 35% of code lines are covered by test cases.
