### 2022-10-21 #534 @emilyriederer
Hi @pachadotdev ! Thanks for starting to dog into the review. Could you please use our review template [here](https://devguide.ropensci.org/reviewtemplate.html#reviewtemplate) to collect your questions and comments? I definitely appreciate the push for testing best practices, but we also want to make sure authors can get the "big picture ideas" out of the reviews
