### 2022-10-25 #534 @pachadotdev
---

1. can you please add vcr to suggests? I would prefer to be able to test offline without skipping the test but having the cassettes to run the test anyways, as in https://github.com/ropensci/tradestatistics/blob/master/tests/testthat/test-ots_create_tidy_data.R#L9

2. do you have a roadmap to test the API-centric functions? these seem to be the most important but here only 35% of code lines are covered by test cases.

1+2 would really simplify the testing for this package

overall it looks good, my only criticism is the lack of a wide battery of tests
