### 2022-10-25 1150c7f
Added vcr to Suggests
The request ir here https://github.com/ropensci/software-review/issues/534#issuecomment-1289832660
