### 2022-10-25 #534 @wilsonfreitas
Hi @pachadotdev

Thanks for your review.
vcr has been added to Suggests.
I believe I don't have a structured roadmap to test the web-scraping functions.
I understand and agree with your suggestions and I'd love to put more work on that, but at this moment I am struggling out of time.
