### 2022-10-28 #534 @quishqa
Dear @emilyriederer , and @wilsonfreitas , sorry once again for the delay.
Here are my comments.

---

`rb3` is a solid package that works and fulfills its objectives. I ran the examples in the README and the vignettes without problems.
I used  `pkgreviewr` to help me review `rb3`. The complete report is [here](https://github.com/quishqa/my_ropensci_reviews/blob/main/rb3-review/rb3-review.md). These are my main comments:

- When running  `goodpractice`, It still detects that the test only covered 35% of the code. And It suggests to used `vapply` instead of `sapply`in `tests/testthat/test-transmute.R`.
- There are many English typos in the documentation (e.g. accross, aditional, compositon, etc). You can see the words to fix [here](https://github.com/quishqa/my_ropensci_reviews/blob/main/rb3-review/rb3-review.md#spell-check).
- When running `test()` there is a warning about [tidyselect](https://github.com/quishqa/my_ropensci_reviews/blob/main/rb3-review/rb3-review.md#run-tests-on-rb3-source): `Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0`.
You can find more information about this warning [here](https://github.com/tidyverse/tidyverse.org/pull/600) and [here](https://community.rstudio.com/t/use-of-data-in-tidyselect-expressions-is-now-deprecated/150092).
- I ran the [Equality data example](https://github.com/wilsonfreitas/rb3#equity-data) without implementing ssl error fix  perfectly fine. I use Ubuntu, maybe it happens in another distro. The test `"it should download a file with a datetime downloader` also ran without problems.
- Also I think it's a good idea to implement GitHub action to regularly check the package.

Congrats for this nice package  !  👍🏽
