### 2022-11-03 #534 @wilsonfreitas
Hi @pachadotdev and @quishqa,

I appreciate your suggestions.
I'll work in the next weeks to get them implemented.
Specially the test coverage for downloads.

Bests,
