### 2022-11-22 #534 @emilyriederer
Hi @wilsonfreitas ! I hope you're well. I just thought I'd check in and see how your updates are going? In general, we try to aim to get an update to our reviewers in ~2-3 weeks after their reviews so they can evaluate any changes while the information is still somewhat fresh in their minds. Thanks!
