### 2022-11-26 0fa1538
rland .data has been deprecated
.data has been deprecated when used with select, arrange, pivot_longer
