### 2022-11-26 7a64d09
Corrected English typos in the documentation
