+++ man/cotahist_get.Rd
@@ -13,8 +13,8 @@ cotahist_get(
 }
 \arguments{
 \item{refdate}{the reference date used to download the file. This reference
-date will be formated as year/month/day according to the given type.
-Accepts ISO formated date strings.}
+date will be formatted as year/month/day according to the given type.
+Accepts ISO formatted date strings.}
 
 \item{type}{a string with \code{yearly} for all data of the given year, \code{monthly}
 for all data of the given month and \code{daily} for the given day.}
@@ -23,7 +23,7 @@ for all data of the given month and \code{daily} for the given day.}
 
 \item{do_cache}{Whether to use cache or not (default = TRUE)
 
-All valueable information is in the \code{HistoricalPrices} element of the
+All valuable information is in the \code{HistoricalPrices} element of the
 returned list.
 \code{Header} and \code{Trailer} have informations regarding file generation.
 The \code{HistoricalPrices} element has a data.frame with data of many assets
