+++ man/rb3-package.Rd
@@ -19,7 +19,7 @@ way the package works and display its alerts.
 rb3 cache folder is named \code{rb3-cache} and it is created inside the
 directory returned by \code{base::tempdir}.
 Since it is changed for every new session it is interesting to use the
-same directory for cache accross sessions.
+same directory for cache across sessions.
 Once the option \code{rb3.cachedir} is set the files are always cached in
 the same directory.
 This is very useful to build a historical data.
