### 2022-11-26 2cce42c
use vapply instead of sapply
