### 2022-11-26 0f3635b
Corrected English typos in the documentation
