+++ inst/extdata/templates/GetTheoricalPortfolio.yaml
@@ -14,24 +14,24 @@ parts:
   Header:
     name: header
     fields:
-    - name: part
-      description: Participação % no índice
-    - name: theoreticalQty
-      description: Quantidade teórica no índice
-    - name: reductor
-      description: Redutor
+      - name: part
+        description: Participação % no índice
+      - name: theoricalQty
+        description: Quantidade teórica no índice
+      - name: reductor
+        description: Redutor
   Results:
     name: results
     fields:
-    - name: code
-      description: Código da ação
-    - name: company
-      description: Nome da companhia
-    - name: spotlight
-      description: Código de especificação da ação
-    - name: theoreticalQty
-      description: Quantidade teórica no índice
-    - name: part
-      description: Participação % no índice
-    - name: cont
-      description: Contagem
+      - name: code
+        description: Código da ação
+      - name: company
+        description: Nome da companhia
+      - name: spotlight
+        description: Código de especificação da ação
+      - name: theoricalQty
+        description: Quantidade teórica no índice
+      - name: part
+        description: Participação % no índice
+      - name: cont
+        description: Contagem
