### 2022-11-26 #534 @wilsonfreitas
Hi @emilyriederer,

I've just pushed changed suggested by @quishqa .

- English typos correction
- use vapply instead of sapply
- removed .data where it is deprecated

I am working now on tests to improve coverage with vcr, as @pachadotdev suggested.
