+++ tests/fixtures/CDIIDI.yml
@@ -1,0 +1,41 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://www2.cetip.com.br/ConsultarTaxaDi/ConsultarTaxaDICetip.aspx
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept: application/json, text/xml, application/xml, */*
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      access-control-allow-headers: Content-Type, Authorization
+      access-control-allow-origin: '*'
+      acesss-control-allow-origin: '*'
+      cache-control: max-age=0, no-cache, no-store
+      connection: keep-alive
+      content-length: '87'
+      content-type: text/html; charset=utf-8
+      date: Thu, 08 Dec 2022 08:03:41 GMT
+      expires: Thu, 08 Dec 2022 08:03:41 GMT
+      pragma: no-cache
+      server: Microsoft-IIS/8.5
+      set-cookie:
+      - CetipLanguage=pt-BR; path=/
+      - ASP.NET_SessionId=3ie0phkegrhm5rw5ouwkhdhq; path=/; HttpOnly
+      strict-transport-security: max-age=15768000 ; includeSubDomains
+      x-content-type-options: nosniff
+      x-frame-options: ALLOW-FROM http://bvmf.bmfbovespa.com.br,http://www2.bmfbovespa.com.br,http://www.bmf.com.br,http://www2.bmf.com.br,https://www.b3.com.br,http://estatisticas.cetip.com.br,https://sistemasweb.b3.com.br/
+      x-powered-by: ASP.NET
+      x-xss-protection: 1; mode=block
+    body:
+      encoding: ''
+      file: no
+      string: '{"taxa":"13,65","dataTaxa":"07/12/2022","indice":"37.558,60","dataIndice":"08/12/2022"}'
+  recorded_at: 2022-12-08 08:03:40 GMT
+  recorded_with: vcr/1.2.0, webmockr/0.8.2
