+++ tests/fixtures/GetListedCashDividends.yml
@@ -1,0 +1,51 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://sistemaswebb3-listados.b3.com.br/listedCompaniesProxy/CompanyCall/GetListedCashDividends/eyJ0cmFkaW5nTmFtZSI6IkFNQkVWU0EiLCJsYW5ndWFnZSI6InB0LWJyIiwicGFnZU51bWJlciI6MSwicGFnZVNpemUiOjk5OTl9
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept: application/json, text/xml, application/xml, */*
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      content-length: '9349'
+      content-type: application/json; charset=utf-8
+      strict-transport-security: max-age=31536000; includeSubDomains
+      x-xss-protection: 1; mode=block
+      x-content-type-options: nosniff
+      content-security-policy: 'default-src ''self'' https://*.google-analytics.com;
+        font-src ''self'' data: https://fonts.gstatic.com/ ;img-src * data: ; script-src
+        ''unsafe-eval'' * ''unsafe-inline''; style-src * ''unsafe-inline''; frame-src
+        http://*.b3.com.br https://*.b3.com.br https://s.tradingview.com/'
+      referrer-policy: strict-origin
+      server-timing: dtSInfo;desc="0", dtRpid;desc="-741119383"
+      date: Thu, 08 Dec 2022 08:36:00 GMT
+      set-cookie: TS01871345=016e3b076f25fd7a5f71b194db8bd7013c49d01a2176bbf2f144de8cb92cf9b7f603b56dabb5bdc386fffc560cbdb549fa5646c528;
+        Path=/; HTTPOnly
+    body:
+      encoding: ''
+      file: no
+      string: '{"page":{"pageNumber":1,"pageSize":9999,"totalRecords":29,"totalPages":1},"results":[{"typeStock":"ON","dateApproval":"09/12/2021","valueCash":"0,1334","ratio":"1","corporateAction":"DIVIDENDO","lastDatePriorEx":"17/12/2021","dateClosingPricePriorExDate":"17/12/2021","closingPricePriorExDate":"16,07","quotedPerShares":"1","corporateActionPrice":"0,830118","lastDateTimePriorEx":"2021-12-17T00:00:00"},{"typeStock":"ON","dateApproval":"09/12/2021","valueCash":"0,4702","ratio":"1","corporateAction":"JRS
+        CAP PROPRIO","lastDatePriorEx":"17/12/2021","dateClosingPricePriorExDate":"17/12/2021","closingPricePriorExDate":"16,07","quotedPerShares":"1","corporateActionPrice":"2,925949","lastDateTimePriorEx":"2021-12-17T00:00:00"},{"typeStock":"ON","dateApproval":"21/12/2020","valueCash":"0,0767","ratio":"1","corporateAction":"DIVIDENDO","lastDatePriorEx":"13/01/2021","dateClosingPricePriorExDate":"13/01/2021","closingPricePriorExDate":"16,17","quotedPerShares":"1","corporateActionPrice":"0,474335","lastDateTimePriorEx":"2021-01-13T00:00:00"},{"typeStock":"ON","dateApproval":"09/12/2020","valueCash":"0,4137","ratio":"1","corporateAction":"JRS
[…]

[…]
