### 2022-12-10 fe77fbf
Added more tests using vcr
