+++ tests/fixtures/FPR.yml
@@ -1,0 +1,646 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://www.b3.com.br/pesquisapregao/download?filelist=FP221207.zip
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept: application/json, text/xml, application/xml, */*
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      cache-control: max-age=2592000
+      content-disposition: filename="pesquisa-pregao.zip"
+      content-encoding: gzip
+      content-security-policy-report-only: 'default-src *.b3.com.br ''self'' blob:
+        data: ''unsafe-inline'' ''unsafe-eval''; img-src *.b3.com.br http://t.co *.google.com
+        *.google.com.br https://*.linkedin.com https://*.adsymptotic.com *.gov.br
+        https://*.rd.services https://*.google-analytics.com https://*.googletagmanager.com
+        data: blob:; style-src *.b3.com.br https://cdnjs.cloudflare.com https://*.googleapis.com
+        data: ''unsafe-inline'' ''unsafe-eval''; font-src *.b3.com.br https://cdnjs.cloudflare.com
+        *.gstatic.com data: ''unsafe-inline'' ''unsafe-eval''; frame-src ''self''
+        *.bmfbovespa.com.br *.bmf.com.br *.b3.com.br *.cetip.com.br https://s.tradingview.com/
+        https://wa-cert-up2dataweb.azurewebsites.net *.cetipmeusinvestimentos.com.br
+        https://www.youtube.com https://*.hotjar.com; connect-src *.b3.com.br https://in.hotjar.com
+        https://stats.g.doubleclick.net https://*.rdstation.com.br https://*.cookielaw.org
+        https://*.google-analytics.com; object-src *.b3.com.br ''self'' ''unsafe-inline''
+        ''unsafe-eval''; script-src *.b3.com.br http://gadgets2.virtualtarget.com.br
[…]
