+++ tests/fixtures/GetTheoricalPortfolio.yml
@@ -1,0 +1,78 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://sistemaswebb3-listados.b3.com.br/indexProxy/indexCall/GetTheoricalPortfolio/eyJwYWdlTnVtYmVyIjoxLCJwYWdlU2l6ZSI6OTk5OSwibGFuZ3VhZ2UiOiJwdC1iciIsImluZGV4IjoiSUJPViJ9
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Accept: application/json, text/xml, application/xml, */*
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      content-length: '10267'
+      content-security-policy: 'default-src ''self'' https://*.google-analytics.com;
+        font-src ''self'' data: https://fonts.gstatic.com/ ;img-src * data: ; script-src
+        ''unsafe-eval'' * ''unsafe-inline''; style-src * ''unsafe-inline''; frame-src
+        http://*.b3.com.br https://*.b3.com.br https://s.tradingview.com/'
+      content-type: application/json; charset=utf-8
+      date: Sat, 10 Dec 2022 09:34:59 GMT
+      referrer-policy: strict-origin
+      server-timing: dtSInfo;desc="0", dtRpid;desc="-1976565348"
+      set-cookie:
+      - dtCookie=v_4_srv_33_sn_E048ED3E9D19F5911B8462AEA4D31FE2_perc_100000_ol_0_mul_1_app-3A2fa0c7805985f6bf_1_rcs-3Acss_0;
+        Path=/; Domain=.b3.com.br
+      - TS01871345=011d592ce158af63b29b820646f49affc204c58e179f65b57325968576243925d251836fcde3c9ec868404bf5a99993b44b0774c8e;
+        Path=/; HTTPOnly
+      - TS0134a800=011d592ce158af63b29b820646f49affc204c58e179f65b57325968576243925d251836fcde3c9ec868404bf5a99993b44b0774c8e;
+        path=/; domain=.b3.com.br; HTTPonly
+      - BIGipServerpool_sistemaswebb3-listados_8443_WAF=1329140746.64288.0000; path=/;
+        Httponly; Secure
+      strict-transport-security: max-age=31536000; includeSubDomains
+      x-content-type-options: nosniff
+      x-xss-protection: 1; mode=block
+    body:
+      encoding: ''
+      file: no
+      string: '{"page":{"pageNumber":1,"pageSize":9999,"totalRecords":92,"totalPages":1},"header":{"part":"100,000","theoricalQty":"99.273.271.525","reductor":"18.842.281,23696071"},"results":[{"cod":"ABEV3","asset":"AMBEV
+        S/A","type":"ON","theoricalQty":"4.386.652.506","part":"3,251","cont":3},{"cod":"ALPA4","asset":"ALPARGATAS","type":"PN      N1","theoricalQty":"202.765.994","part":"0,202","cont":2},{"cod":"AMER3","asset":"AMERICANAS","type":"ON      NM","theoricalQty":"596.086.291","part":"0,443","cont":4},{"cod":"ARZZ3","asset":"AREZZO
[…]
