### 2022-12-11 #534 @pachadotdev
> Hi @pachadotdev,
>
> I added vcr to the tests. I'm not finished yet. I'd appreciate it if you could give me any suggestions, especially to understand that I'm on the correct path.
>
> This [commit](76c13e795fb5d5c2c27550bdd0f63341911cf476) has the vcr implementation.
>
> I believe I'll have it finished this weekend.
>
> thank you so much for your attention and participation.

Awesome!!! I have to present thesis progress this week, so I'll be quite disconnected. After that, I can give a careful look.

Right now, I am redeploying my server, because I have over 700 GB of data, and I needed a cheaper VPS.

Best,
