### 2022-12-12 #534 @wilsonfreitas
Great @pachadotdev,

I've implemented more tests and I believe we can go on with the review process.
