### 2022-12-26 #534 @emilyriederer
Hi @wilsonfreitas ! Thanks for the heads up. Just to confirm: does that mean you're ready for the reviewers to look again for their final approval?
