### 2022-12-28 #534 @emilyriederer
Hi there @pachadotdev and @quishqa ! Could you please take a look an @wilsonfreitas 's changes and complete the [approval template](https://devguide.ropensci.org/reviewerguide.html#followupreviewer)? Thank you!
