### 2023-01-19 #534 @quishqa
Hi all!
I believe the commits addressed my suggestions, but it seems that [R-CMD-Check is failing](https://github.com/wilsonfreitas/rb3/actions/runs/3664411763).
