### 2023-01-20 #534 @wilsonfreitas
Hi @quishqa, I forgot to update the GitHub workflows.
The bizdays package has also been updated, and it is a dependency.
It's submitted to CRAN, I'll wait until it's approved to rerun the workflows.
