### 2023-01-20 #534 @pachadotdev
> Hi @pachadotdev and @quishqa - happy new year! I know my last message might have been poorly timed around the holidays. Would you have time soon to do a final review and approval of @wilsonfreitas 's updates? Thank you!

hi!
looks good!
