### 2023-01-21 #534 @emilyriederer
Thanks for checking in @quishqa and @pachadotdev ! Once @wilsonfreitas has the checks back, can you both please formally sign off with the [template](https://devguide.ropensci.org/reviewerguide.html#followupreviewer)?
