### 2023-01-22 339613b
use an explicit call to `as.character()` within `map_chr()`
Automatic coercion from double to character was deprecated in purrr 1.0.0.
