+++ tests/testthat/setup.R
@@ -1,4 +1,3 @@
-
 local_cachedir <- file.path(tempdir(), "rb3-cache")
 withr::defer(unlink(local_cachedir, recursive = TRUE), teardown_env())
 
@@ -6,4 +5,5 @@ op <- options(
   rb3.cachedir = local_cachedir,
   rb3.silent = TRUE
 )
+
 withr::defer(options(op), teardown_env())
