### 2023-01-24 #534 @wilsonfreitas
Hi @pachadotdev and @quishqa, all [RMD checks](https://github.com/wilsonfreitas/rb3/actions/runs/3990979812) are passing.
