### 2023-01-25 #534 @quishqa
> Thanks for checking in @quishqa and @pachadotdev ! Once @wilsonfreitas has the checks back, can you both please formally sign off with the [template](https://devguide.ropensci.org/reviewerguide.html#followupreviewer)?

Just double-check [this is the template](https://devguide.ropensci.org/approval2template.html), right?
