### 2023-02-08 #534 @emilyriederer
Congratulations, @wilsonfreitas ! Yuur package has been approved. There's a lot of great information above about how you can fully transfer the package to rOpenSci and some (optional) ideas for marketing.

Would you also be interested in writing a blog post about your package? I think it's always great to feature new data packages in posts since it can raise awareness both for the package and for new underlying APIs that folks may not know about. Additionally, I think there are a lot of synergies between this package and other @msperlin and @pachadotdev submissions, so we could even explore those interactions. If so, @ropensci/blog-editors can provide more detail.
