### 2023-02-17 #534 @wilsonfreitas
hi @emilyriederer

I am trying to transfer the repo to ropensci, but I don't have the permission.

![image](https://user-images.githubusercontent.com/1583922/219600754-ec596034-144b-4bf6-af20-91b572c786c9.png)

Am I doing it correctly?
