### 2023-02-17 #534 @maelle
@wilsonfreitas I've now invited you to the GitHub organization, can you please try again?

I've just realized this submission is a scenario we hadn't thought of: the submitter @msperlin didn't own the repository, and we send the organization invitation to the submitter. :sweat_smile: Sorry about the resulting clunkiness!
